### 2022-02-27 #4209 @arfon
@althonos – this PR should fix your paper compilation: https://github.com/althonos/pyrodigal/pull/9
