### 2022-02-27 60aa0c2
Fix affiliation in paper metadata (#9)
