### 2022-02-27 #4209 @althonos
Thank you!

> @althonos if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Given their experience with Python+C and their knowledge of genomics, any I suppose the following reviewers would be a good fit:
- @\chaypaterson
- @\Gab0
- @\standage
- @\will-rowe
