### 2022-03-11 1e8eaae
Update `add_nodes` function to support custom minimum gene length
