### 2022-03-11 d30c8f8
Refactor the `add_nodes` function into a method of `Nodes` objects
