+++ pyrodigal/tests/test_nodes.py
@@ -5,7 +5,7 @@ import sys
 import unittest
 
 from .. import Nodes, Sequence
-from .._pyrodigal import METAGENOMIC_BINS, add_nodes
+from .._pyrodigal import METAGENOMIC_BINS
 
 from .fasta import parse
 
@@ -26,23 +26,23 @@ class TestNodes(unittest.TestCase):
         self.assertEqual(len(nodes), 0)
         # numbers below obtained directly in Prodigal by `printf`-ing the
         # node numbers on a normal run
-        self.assertEqual(add_nodes(nodes, seq, METAGENOMIC_BINS[0].training_info), 2970)
+        self.assertEqual(nodes.extract(seq, METAGENOMIC_BINS[0].training_info), 2970)
         self.assertEqual(len(nodes), 2970)
         nodes.clear()
-        self.assertEqual(add_nodes(nodes, seq, METAGENOMIC_BINS[2].training_info), 2970)
+        self.assertEqual(nodes.extract(seq, METAGENOMIC_BINS[2].training_info), 2970)
         self.assertEqual(len(nodes), 2970)
         nodes.clear()
-        self.assertEqual(add_nodes(nodes, seq, METAGENOMIC_BINS[11].training_info), 2293)
+        self.assertEqual(nodes.extract(seq, METAGENOMIC_BINS[11].training_info), 2293)
         self.assertEqual(len(nodes), 2293)
         nodes.clear()
-        self.assertEqual(add_nodes(nodes, seq, METAGENOMIC_BINS[24].training_info), 2293)
+        self.assertEqual(nodes.extract(seq, METAGENOMIC_BINS[24].training_info), 2293)
         self.assertEqual(len(nodes), 2293)
         nodes.clear()
 
     def test_copy(self):
         seq = Sequence.from_string(self.record.seq)
         nodes1 = Nodes()
-        add_nodes(nodes1, seq, METAGENOMIC_BINS[0].training_info)
+        nodes1.extract(seq, METAGENOMIC_BINS[0].training_info)
         nodes2 = nodes1.copy()
         for n1, n2 in zip(nodes1, nodes2):
             self.assertEqual(n1.type, n2.type)
