+++ pyrodigal/_pyrodigal.pyx
@@ -859,6 +859,8 @@ cdef class Nodes:
 
     """
 
+    # --- Magic methods ------------------------------------------------------
+
     def __cinit__(self):
         self.nodes = NULL
         self.capacity = 0
@@ -889,6 +891,8 @@ cdef class Nodes:
 
     def __sizeof__(self):
         return self.capacity * sizeof(_node) + sizeof(self)
+
+    # --- C interface --------------------------------------------------------
 
     cdef inline _node* _add_node(
         self,
@@ -928,13 +932,199 @@ cdef class Nodes:
         old_length, self.length = self.length, 0
         memset(self.nodes, 0, old_length * sizeof(_node))
 
-    def clear(self):
-        """Remove all nodes from the vector.
+    cdef int _extract(
+        self,
+        Sequence sequence,
+        TrainingInfo training_info,
+        bint closed=False,
+        int min_gene=MIN_GENE,
+        int min_edge_gene=MIN_EDGE_GENE,
+    ) nogil except -1:
+        cdef int    i
+        cdef int    last[3]
+        cdef int    min_dist[3]
+        cdef bint   saw_start[3]
+        cdef int    slmod        = sequence.slen % 3
+        cdef int    nn           = 0
+        cdef int    tt           = training_info.tinf.trans_table
+        cdef _mask* mlist        = sequence.masks.masks
+        cdef int    nm           = sequence.masks.length
+
+        # If sequence is smaller than a codon, there are no nodes to add
+        if sequence.slen < 3:
+            return nn
+
+        # Forward strand nodes
+        for i in range(3):
+            last[(i+slmod)%3] = sequence.slen + i
+            saw_start[i%3] = False
+            min_dist[i%3] = min_edge_gene
+            if not closed:
+                while last[(i+slmod)%3] + 3 > sequence.slen:
+                    last[(i+slmod)%3] -= 3
+        for i in reversed(range(sequence.slen-2)):
+            if sequence._is_stop(i, tt):
+                if saw_start[i%3]:
+                    self._add_node(
+                        ndx = last[i%3],
+                        type = node_type.STOP,
+                        strand = 1,
+                        stop_val = i,
+                        edge = not sequence._is_stop(last[i%3], tt),
+                    )
+                    nn += 1
+                min_dist[i%3] = min_gene
+                last[i%3] = i
+                saw_start[i%3] = False
+                continue
+            if last[i%3] >= sequence.slen:
+                continue
+            if not cross_mask(i, last[i%3], mlist, nm):
+                if last[i%3] - i + 3 >= min_dist[i%3] and sequence._is_start(i, tt):
+                    if sequence._is_atg(i):
+                        saw_start[i%3] = True
+                        self._add_node(
+                            ndx = i,
+                            type = node_type.ATG,
+                            stop_val = last[i%3],
+                            strand = 1,
+                            edge = False
+                        )
+                        nn += 1
+                    elif sequence._is_ttg(i):
+                        saw_start[i%3] = True
+                        self._add_node(
+                            ndx = i,
+                            type = node_type.TTG,
+                            stop_val = last[i%3],
+                            strand = 1,
+                            edge = False
+                        )
+                        nn += 1
+                    elif sequence._is_gtg(i):
+                        saw_start[i%3] = True
+                        self._add_node(
+                            ndx = i,
+                            type = node_type.GTG,
+                            stop_val = last[i%3],
+                            strand = 1,
+                            edge = False
+                        )
+                        nn += 1
+                if i <= 2 and not closed and last[i%3] - i > min_edge_gene:
+                    saw_start[i%3] = True
[…]
