### 2022-03-11 829a31c
Refactor `OrfFinder.train` to support custom minimum gene size
