+++ pyrodigal/_pyrodigal.pyx
@@ -2062,21 +2062,31 @@ cdef class OrfFinder:
 
     """
 
+    # --- Magic methods ------------------------------------------------------
+
     def __cinit__(self):
         self._num_seq = 1
 
     def __init__(
         self,
+        TrainingInfo training_info=None,
+        *,
         bint meta=False,
         bint closed=False,
         bint mask=False,
-        TrainingInfo training_info=None,
+        int min_gene=MIN_GENE,
+        int min_edge_gene=MIN_EDGE_GENE,
     ):
-        """__init__(self, meta=False, closed=False, mask=False)\n--
+        f"""__init__(self, training_info=None, *, meta=False, closed=False, mask=False, min_gene={MIN_GENE}, min_edge_gene={MIN_EDGE_GENE})\n--
 
         Instantiate and configure a new ORF finder.
 
         Arguments:
+            training_info (`~pyrodigal.TrainingInfo`, optional): A training
+                info instance to use in single mode without having to
+                train first.
+
+        Keyword Arguments:
             meta (`bool`): Set to `True` to run in metagenomic mode, using
                 a pre-trained profiles for better results with metagenomic
                 or progenomic inputs. Defaults to `False`.
@@ -2085,9 +2095,10 @@ cdef class OrfFinder:
                 Defaults to `False`.
             mask (`bool`): Prevent genes from running across regions
                 containing unknown nucleotides. Defaults to `False`.
-            training_info (`~pyrodigal.TrainingInfo`, optional): A training
-                info instance to use in single mode without having to
-                train first.
+            min_gene (`int`): The minimum gene length. Defaults to the value
+                used in Prodigal.
+            min_edge_gene (`int`): The minimum edge gene length. Defaults to
+                the value used in Prodigal.
 
         .. versionchanged:: 0.6.4
            Added the ``training_info`` argument.
@@ -2096,11 +2107,18 @@ cdef class OrfFinder:
         if meta and training_info is not None:
             raise ValueError("cannot use a training info in meta mode.")
 
+        if min_gene <= 0:
+            raise ValueError("`min_gene` must be strictly positive")
+        if min_edge_gene <= 0:
+            raise ValueError("`min_edge_gene` must be strictly positive")
+
         self.meta = meta
         self.closed = closed
         self.lock = threading.Lock()
         self.mask = mask
         self.training_info = training_info
+        self.min_gene = min_gene
+        self.min_edge_gene = min_edge_gene
 
     def __repr__(self):
         ty = type(self).__name__
@@ -2119,6 +2137,56 @@ cdef class OrfFinder:
                 self.mask,
                 self.training_info
             )
+
+    # --- C interface --------------------------------------------------------
+
+    cdef int _train(
+        self,
+        Sequence sequence,
+        Nodes nodes,
+        ConnectionScorer scorer,
+        TrainingInfo tinf,
+        bint force_nonsd=False,
+        double start_weight=4.35,
+        int translation_table=11,
+    ) nogil except 1:
+        # find all the potential starts and stops
+        nodes._extract(
+            sequence,
+            tinf,
+            closed=self.closed,
+            min_gene=self.min_gene,
+            min_edge_gene=self.min_edge_gene
+        )
+        nodes._sort()
+        scorer._index(nodes)
+        # scan all the ORFs looking for a potential GC bias in a particular
+        # codon position, in order to acquire a good initial set of genes
+        gc_frame = calc_most_gc_frame(sequence)
+        if not gc_frame:
+            raise MemoryError()
+        node.record_gc_bias(gc_frame, nodes.nodes, nodes.length, tinf.tinf)
+        free(gc_frame)
[…]
