### 2022-03-11 0dd6c2e
Move back `find_genes_single` and `find_genes_meta` as `OrfFinder` methods
