+++ pyrodigal/_pyrodigal.pyx
@@ -2149,7 +2149,7 @@ cdef class OrfFinder:
         bint force_nonsd=False,
         double start_weight=4.35,
         int translation_table=11,
-    ) nogil except 1:
+    ) nogil except -1:
         # find all the potential starts and stops
         nodes._extract(
             sequence,
@@ -2169,7 +2169,7 @@ cdef class OrfFinder:
         free(gc_frame)
         # do an initial dynamic programming routine with just the GC frame bias
         # used as a scoring function.
-        record_overlapping_starts(nodes, tinf, is_meta=False)
+        record_overlapping_starts(nodes, tinf, final=False)
         ipath = dynamic_programming(nodes, tinf, scorer, final=False)
         # gather dicodon statistics for the training set
         calc_dicodon_gene(tinf, sequence, nodes, ipath)
@@ -2185,6 +2185,124 @@ cdef class OrfFinder:
             train_starts_nonsd(nodes, sequence, tinf)
         # return 0 on success
         return 0
+
+    cdef int _find_genes_single(
+        self,
+        Sequence sequence,
+        TrainingInfo tinf,
+        ConnectionScorer scorer,
+        Nodes nodes,
+        Genes genes,
+        int sequence_index
+    ) nogil except -1:
+        cdef int ipath
+        # find all the potential starts and stops, and sort them
+        nodes._extract(
+            sequence,
+            tinf,
+            closed=self.closed,
+            min_gene=self.min_gene,
+            min_edge_gene=self.min_edge_gene
+        )
+        nodes._sort()
+        scorer._index(nodes)
+        # second dynamic programming, using the dicodon statistics as the
+        # scoring function
+        reset_node_scores(nodes)
+        score_nodes(nodes, sequence, tinf, closed=self.closed, is_meta=False)
+        record_overlapping_starts(nodes, tinf, final=True)
+        ipath = dynamic_programming(nodes, tinf, scorer, final=True)
+        # eliminate eventual bad genes in the nodes
+        if nodes.length > 0:
+            eliminate_bad_genes(nodes, ipath, tinf)
+        # record genes
+        add_genes(genes, nodes, ipath)
+        tweak_final_starts(genes, nodes, tinf)
+        record_gene_data(genes, nodes, tinf, sequence_index=sequence_index)
+        # return 0 on success
+        return 0
+
+    cdef int _find_genes_meta(
+        self,
+        Sequence sequence,
+        ConnectionScorer scorer,
+        Nodes nodes,
+        Genes genes,
+        int sequence_index
+    ) except -1:
+        cdef int          i
+        cdef double       low
+        cdef double       high
+        cdef int          ipath
+        cdef TrainingInfo tinf
+        cdef int          tt        = -1
+        cdef int          max_phase = 0
+        cdef double       max_score = -100.0
+
+
+        # compute the min/max acceptable gc for the sequence to only
+        # use appropriate metagenomic bins
+        low = 0.88495*sequence.gc - 0.0102337
+        high = 0.86596*sequence.gc + 0.1131991
+        if low > 0.65:
+            low = 0.65
+        if high < 0.35:
+            high = 0.35
+
+        # check which of the metagenomic bins gets the best results
+        for i in range(NUM_META):
+            # check which of the metagenomic bins gets the best results
+            if _METAGENOMIC_BINS[i].tinf.gc < low or _METAGENOMIC_BINS[i].tinf.gc > high:
+                continue
+            # record the training information for the current bin
+            tinf = METAGENOMIC_BINS[i].training_info
+            # recreate the node list if the translation table changed
+            if tinf.tinf.trans_table != tt:
+                tt = tinf.tinf.trans_table
+                nodes._clear()
+                nodes._extract(
[…]
