### 2022-03-11 d0037f7
Make `Nodes.extract` only take the translation table as argument
