### 2022-03-11 28ef050
Turn `calc_orf_gc` and `raw_coding_score` into `Nodes` methods
