+++ pyrodigal/_pyrodigal.pyx
@@ -925,6 +925,47 @@ cdef class Nodes:
         node.edge = edge
         return node
 
+    cdef int _calc_orf_gc(self, Sequence seq, TrainingInfo tinf) nogil except -1:
+        cdef int i
+        cdef int j
+        cdef int last[3]
+        cdef int phase
+        cdef double gc[3]
+        cdef double gsize = 0.0
+
+        # direct strand
+        gc[0] = gc[1] = gc[2] = 0.0
+        for i in reversed(range(<int> self.length)):
+            if self.nodes[i].strand == 1:
+                phase = self.nodes[i].ndx %3
+                if self.nodes[i].type == node_type.STOP:
+                    last[phase] = j = self.nodes[i].ndx
+                    gc[phase] = seq._is_gc(j) + seq._is_gc(j+1) + seq._is_gc(j+2)
+                else:
+                    for j in range(last[phase] - 3, self.nodes[i].ndx - 1, -3):
+                        gc[phase] += seq._is_gc(j) + seq._is_gc(j+1) + seq._is_gc(j+2)
+                    gsize = abs(self.nodes[i].stop_val - self.nodes[i].ndx) + 3.0
+                    self.nodes[i].gc_cont = gc[phase] / gsize
+                    last[phase] = self.nodes[i].ndx
+
+        # reverse strand
+        gc[0] = gc[1] = gc[2] = 0.0
+        for i in range(<int> self.length):
+            if self.nodes[i].strand == -1:
+                phase = self.nodes[i].ndx % 3
+                if self.nodes[i].type == node_type.STOP:
+                    last[phase] = j = self.nodes[i].ndx
+                    gc[phase] = seq._is_gc(j) + seq._is_gc(j-1) + seq._is_gc(j-2)
+                else:
+                    for j in range(last[phase] + 3, self.nodes[i].ndx + 1, 3):
+                        gc[phase] += seq._is_gc(j) + seq._is_gc(j+1) + seq._is_gc(j+2)
+                    gsize = abs(self.nodes[i].stop_val - self.nodes[i].ndx) + 3.0
+                    self.nodes[i].gc_cont = gc[phase] / gsize
+                    last[phase] = self.nodes[i].ndx
+
+        # return 0 on success
+        return 0
+
     cdef int _clear(self) nogil except 1:
         """Remove all nodes from the vector.
         """
@@ -1111,6 +1152,274 @@ cdef class Nodes:
 
         return nn
 
+    cdef int _raw_coding_score(self, Sequence seq, TrainingInfo tinf) nogil except -1:
+        cdef double  score[3]
+        cdef double  lfac
+        cdef double  no_stop
+        cdef double  gsize
+        cdef ssize_t last[3]
+        cdef size_t  phase
+        cdef ssize_t j
+        cdef ssize_t i
+        cdef ssize_t nn = self.length
+
+        if tinf.tinf.trans_table != 11:
+            no_stop =  ((1-tinf.tinf.gc)*(1-tinf.tinf.gc)*tinf.tinf.gc)     / 8.0
+            no_stop += ((1-tinf.tinf.gc)*(1-tinf.tinf.gc)*(1-tinf.tinf.gc)) / 8.0
+            no_stop = 1 - no_stop
+        else:
+            no_stop =  ((1-tinf.tinf.gc)*(1-tinf.tinf.gc)*tinf.tinf.gc)     / 4.0
+            no_stop += ((1-tinf.tinf.gc)*(1-tinf.tinf.gc)*(1-tinf.tinf.gc)) / 8.0
+            no_stop = 1 - no_stop
+
+        # Initial Pass: Score coding potential (start->stop)
+        score[0] = score[1] = score[2] = 0.0
+        for i in reversed(range(nn)):
[…]
