### 2022-03-11 8bbd7df
Turn `find_best_upstream_motif` into a `Node` static method
