+++ pyrodigal/_pyrodigal.pyx
@@ -810,6 +810,8 @@ cdef class Node:
 
     """
 
+    # --- Properties ---------------------------------------------------------
+
     @property
     def type(self):
         """`str`: The node type (ATG, GTG, TTG, or Stop).
@@ -851,6 +853,73 @@ cdef class Node:
         """
         assert self.node != NULL
         return self.node.score
+
+    # --- C interface --------------------------------------------------------
+
+    @staticmethod
+    cdef int _find_best_upstream_motif(
+        _node* node,
+        Sequence seq,
+        _training* tinf,
+        int stage
+    ) nogil except -1:
+        cdef int i
+        cdef int j
+        cdef int start
+        cdef int spacer
+        cdef int spacendx
+        cdef int index
+        cdef int max_spacer   = 0
+        cdef int max_spacendx = 0
+        cdef int max_len      = 0
+        cdef int max_ndx      = 0
+        cdef double max_sc    = -100.0
+        cdef double score     = 0.0
+
+        if node.type == node_type.STOP or node.edge:
+            return 0
+
+        if node.strand == 1:
+            start = node.ndx
+        else:
+            start = seq.slen - 1 - node.ndx
+
+        for i in reversed(range(4)):
+            for j in range(start-18-i, start-5-i):
+                if j < 0:
+                    continue
+                spacer = start - j - i - 3
+
+                if j <= start - 16 - i:
+                    spacendx = 3
+                elif j <= start - 14 - i:
+                    spacendx = 2
+                elif j >= start - 7 - i:
+                    spacendx = 1
+                else:
+                    spacendx = 0
+
+                index = seq._mer_ndx(j, length=i+3, strand=node.strand)
+                score = tinf.mot_wt[i][spacendx][index]
+                if score > max_sc:
+                    max_sc = score
+                    max_spacendx = spacendx
+                    max_spacer = spacer
+                    max_ndx = index
+                    max_len = i+3
+
+        if stage == 2 and (max_sc == -4.0 or max_sc < tinf.no_mot + 0.69):
+            node.mot.ndx = 0
+            node.mot.len = 0
+            node.mot.spacendx = 0
+            node.mot.spacer = 0
+            node.mot.score = tinf.no_mot
+        else:
+            node.mot.ndx = max_ndx
+            node.mot.len = max_len
+            node.mot.spacendx = max_spacendx
+            node.mot.spacer = max_spacer
+            node.mot.score = max_sc
 
 cdef class Nodes:
     """A list of dynamic programming nodes used by Prodigal to score ORFs.
@@ -925,7 +994,7 @@ cdef class Nodes:
         node.edge = edge
         return node
 
-    cdef int _calc_orf_gc(self, Sequence seq, TrainingInfo tinf) nogil except -1:
+    cdef int _calc_orf_gc(self, Sequence seq) nogil except -1:
         cdef int i
         cdef int j
         cdef int last[3]
@@ -1152,7 +1221,7 @@ cdef class Nodes:
 
         return nn
 
-    cdef int _raw_coding_score(self, Sequence seq, TrainingInfo tinf) nogil except -1:
+    cdef int _raw_coding_score(self, Sequence seq, _training* tinf) nogil except -1:
         cdef double  score[3]
         cdef double  lfac
         cdef double  no_stop
@@ -1163,13 +1232,13 @@ cdef class Nodes:
         cdef ssize_t i
         cdef ssize_t nn = self.length
 
-        if tinf.tinf.trans_table != 11:
[…]
