### 2022-03-11 c0b9fc0
Make `shine_dalgarno_mm` and `shine_dalgarno_exact` methods of `Sequence` objects
