+++ pyrodigal/_pyrodigal.pyx
@@ -240,29 +240,7 @@ cdef class Sequence:
 
     """
 
-    def __cinit__(self):
-        self.slen = 0
-        self.gc = 0.0
-        self.digits = NULL
-        self.masks = None
-
-    def __dealloc__(self):
-        PyMem_Free(self.digits)
-
-    def __len__(self):
-        return self.slen
-
-    def __sizeof__(self):
-        return self.slen * sizeof(uint8_t) + sizeof(self)
-
-    cdef int _allocate(self, int slen) except 1:
-        self.slen = slen
-        self.digits = <uint8_t*> PyMem_Malloc(slen * sizeof(uint8_t))
-        if self.digits == NULL:
-            raise MemoryError()
-        with nogil:
-            memset(self.digits, 0, slen * sizeof(uint8_t))
-        return 0
+    # --- Class methods ------------------------------------------------------
 
     @classmethod
     def from_bytes(cls, const unsigned char[:] sequence, bint mask = False):
@@ -382,6 +360,34 @@ cdef class Sequence:
                             mask_begin = -1
 
         return seq
+
+    # --- Magic methods ------------------------------------------------------
+
+    def __cinit__(self):
+        self.slen = 0
+        self.gc = 0.0
+        self.digits = NULL
+        self.masks = None
+
+    def __dealloc__(self):
+        PyMem_Free(self.digits)
+
+    def __len__(self):
+        return self.slen
+
+    def __sizeof__(self):
+        return self.slen * sizeof(uint8_t) + sizeof(self)
+
+    # --- C interface -------------------------------------------------------
+
+    cdef int _allocate(self, int slen) except 1:
+        self.slen = slen
+        self.digits = <uint8_t*> PyMem_Malloc(slen * sizeof(uint8_t))
+        if self.digits == NULL:
+            raise MemoryError()
+        with nogil:
+            memset(self.digits, 0, slen * sizeof(uint8_t))
+        return 0
 
     cdef inline bint _is_a(self, int i, int strand = 1) nogil:
         cdef uint8_t x
@@ -629,6 +635,217 @@ cdef class Sequence:
             return b"G"
 
         return b'X'
+
+    cdef int _shine_dalgarno_exact(self, int pos, int start, _training* tinf, int strand=1) nogil:
+        cdef int i
+        cdef int j
+        cdef int k
+        cdef int mism
+        cdef int rdis
+        cdef int limit
+        cdef int max_val
+        cdef int cmp_val
+        cdef int cur_val = 0
+        cdef int match[6]
+        cdef int cur_ctr
+        cdef int dis_flag
+
+        limit = start - 4 - pos
+        if limit > 6:
+            limit = 6
+        for i in range(limit, 6):
+            match[i] = -10
+
+        # Compare the 6-base region to AGGAGG
+        for i in range(limit):
+            if pos + i < 0:
+                continue
+            if i%3 == 0 and self._is_a(pos+i, strand=strand):
+                match[i] = 2
+            elif i%3 != 0 and self._is_g(pos+i, strand=strand):
+                match[i] = 3
+            else:
+                match[i] = -10
+
+        # Find the maximally scoring motif
+        max_val = 0
+        for i in range(limit, 2, -1):
+            for j in range(limit+1-i):
+                cur_ctr = -2
+                mism = 0;
+                for k in range(j, j+i):
+                    cur_ctr += match[k]
+                    if match[k] < 0:
+                        mism += 1
+                if mism > 0:
+                    continue
+                rdis = start - (pos + j + i)
[…]
