### 2022-03-11 e5353cc
Make `Nodes._find_genes_meta` work without having to hold the GIL
