+++ pyrodigal/_pyrodigal.pyx
@@ -1000,14 +1000,12 @@ cdef class ConnectionScorer:
         with nogil:
             self._compute_skippable(min, i)
 
-    cdef int _score_connections(self, Nodes nodes, int min, int i, TrainingInfo tinf, bint final=False) nogil except 1:
+    cdef int _score_connections(self, Nodes nodes, int min, int i, _training* tinf, bint final=False) nogil except 1:
         cdef int j
-        cdef _node*     raw_nodes = nodes.nodes
-        cdef _training* raw_tinf  = tinf.tinf
 
         for j in range(min, i):
             if self.skip_connection[j] == 0:
-                dprog.score_connection(raw_nodes, j, i, raw_tinf, final)
+                dprog.score_connection(nodes.nodes, j, i, tinf, final)
 
         return 0
 
@@ -1016,7 +1014,7 @@ cdef class ConnectionScorer:
         assert i < <int> nodes.length
         assert min <= i
         with nogil:
-            self._score_connections(nodes, min, i, tinf, final)
+            self._score_connections(nodes, min, i, tinf.tinf, final)
 
 # --- Nodes ------------------------------------------------------------------
 
@@ -1074,12 +1072,12 @@ cdef class Node:
     # --- C interface --------------------------------------------------------
 
     @staticmethod
-    cdef int _find_best_upstream_motif(
+    cdef void _find_best_upstream_motif(
         _node* node,
         Sequence seq,
         _training* tinf,
         int stage
-    ) nogil except -1:
+    ) nogil:
         cdef int i
         cdef int j
         cdef int start
@@ -1094,7 +1092,7 @@ cdef class Node:
         cdef double score     = 0.0
 
         if node.type == node_type.STOP or node.edge:
-            return 0
+            return
 
         if node.strand == 1:
             start = node.ndx
@@ -1137,6 +1135,41 @@ cdef class Node:
             node.mot.spacendx = max_spacendx
             node.mot.spacer = max_spacer
             node.mot.score = max_sc
+
+    @staticmethod
+    cdef void _score_upstream_composition(
+        _node* node,
+        Sequence seq,
+        _training* tinf,
+    ) nogil:
+        cdef int i
+        cdef int start
+        cdef int count = 0
+        cdef int mer
+        cdef int strand
+
+        if node.strand == 1:
+            start = node.ndx
+            strand = 1
+        else:
+            start = seq.slen - 1 - node.ndx
+            strand = -1
+
+        node.uscore = 0.0;
+        for i in range(1, 45):
+            if i > 2 and i < 15:
+                continue
+            if start < i:
+                continue
+
+            if strand == 1:
+                mer = seq.digits[start - i] & 0b11
+            else:
+                mer = _translation[seq.digits[seq.slen - 1 - start + i]] & 0b11
+
+            node.uscore += 0.4 * tinf.st_wt * tinf.ups_comp[count][mer]
+            count += 1
+
 
 cdef class Nodes:
     """A list of dynamic programming nodes used by Prodigal to score ORFs.
@@ -1258,6 +1291,90 @@ cdef class Nodes:
         cdef size_t old_length
         old_length, self.length = self.length, 0
         memset(self.nodes, 0, old_length * sizeof(_node))
+
+    cdef int _dynamic_programming(self, _training* tinf, ConnectionScorer scorer, bint final=False) nogil:
+        cdef int    i
+        cdef int    j
+        cdef int    min
+        cdef int    path
+        cdef int    nxt
+        cdef int    tmp
+        cdef int    max_ndx = -1
+        cdef double max_sc  = -1.0
+
+        if self.length == 0:
+            return -1
+
[…]
