### 2022-03-11 532e7cd
Turn `calc_dicodon_gene` into a `TrainingInfo` method
