+++ pyrodigal/_pyrodigal.pyx
@@ -204,6 +204,7 @@ cdef class Masks:
         memcpy(new.masks, self.masks, new.capacity * sizeof(_mask))
         return new
 
+
 # --- Input sequence ---------------------------------------------------------
 
 cdef enum:
@@ -213,13 +214,13 @@ cdef enum:
     T = 0b011
     N = 0b110  # use 6 so that N & 0x3 == C
 
-cdef uint8_t _translation[N+1]
+cdef uint8_t _complement[N+1]
 for i in range(N+1):
-    _translation[i] = N
-_translation[A] = T
-_translation[T] = A
-_translation[C] = G
-_translation[G] = C
+    _complement[i] = N
+_complement[A] = T
+_complement[T] = A
+_complement[C] = G
+_complement[G] = C
 
 cdef Py_UCS4 _letters[N+1]
 for i in range(N+1):
@@ -228,7 +229,6 @@ _letters[A] = "A"
 _letters[T] = "T"
 _letters[C] = "C"
 _letters[G] = "G"
-
 
 cdef class Sequence:
     """A digitized input sequence.
@@ -430,9 +430,9 @@ cdef class Sequence:
             x1 = self.digits[i+1]
             x2 = self.digits[i+2]
         else:
-            x0 = _translation[self.digits[self.slen - 1 - i]]
-            x1 = _translation[self.digits[self.slen - 2 - i]]
-            x2 = _translation[self.digits[self.slen - 3 - i]]
+            x0 = _complement[self.digits[self.slen - 1 - i]]
+            x1 = _complement[self.digits[self.slen - 2 - i]]
+            x2 = _complement[self.digits[self.slen - 3 - i]]
 
         # ATG
         if x0 == A and x1 == T and x2 == G:
@@ -459,9 +459,9 @@ cdef class Sequence:
             x1 = self.digits[i+1]
             x2 = self.digits[i+2]
         else:
-            x0 = _translation[self.digits[self.slen - 1 - i]]
-            x1 = _translation[self.digits[self.slen - 2 - i]]
-            x2 = _translation[self.digits[self.slen - 3 - i]]
+            x0 = _complement[self.digits[self.slen - 1 - i]]
+            x1 = _complement[self.digits[self.slen - 2 - i]]
+            x2 = _complement[self.digits[self.slen - 3 - i]]
 
         # TAG
         if x0 == T and x1 == A and x2 == G:
@@ -497,9 +497,9 @@ cdef class Sequence:
             x1 = self.digits[i+1]
             x2 = self.digits[i+2]
         else:
-            x0 = _translation[self.digits[self.slen - 1 - i]]
-            x1 = _translation[self.digits[self.slen - 2 - i]]
-            x2 = _translation[self.digits[self.slen - 3 - i]]
+            x0 = _complement[self.digits[self.slen - 1 - i]]
+            x1 = _complement[self.digits[self.slen - 2 - i]]
+            x2 = _complement[self.digits[self.slen - 3 - i]]
         return x0 == A and x1 == T and x2 == G
 
     cdef inline bint _is_gtg(self, int i, int strand = 1) nogil:
@@ -511,9 +511,9 @@ cdef class Sequence:
             x1 = self.digits[i+1]
             x2 = self.digits[i+2]
         else:
-            x0 = _translation[self.digits[self.slen - 1 - i]]
-            x1 = _translation[self.digits[self.slen - 2 - i]]
-            x2 = _translation[self.digits[self.slen - 3 - i]]
+            x0 = _complement[self.digits[self.slen - 1 - i]]
+            x1 = _complement[self.digits[self.slen - 2 - i]]
[…]
