### 2022-03-11 5bf9b0e
Continue refactoring some training functions into `TrainingInfo`
