+++ pyrodigal/_pyrodigal.pyx
@@ -2273,6 +2273,64 @@ cdef class TrainingInfo:
 
     # --- C interface --------------------------------------------------------
 
+    @staticmethod
+    cdef void _update_motif_counts(double mcnt[4][4][4096], double *zero, Sequence seq, _node* nod, int stage) nogil:
+        cdef int     i
+        cdef int     j
+        cdef int     k
+        cdef int     start
+        cdef int     spacendx
+        cdef _motif* mot      = &nod.mot
+
+        if nod.type == node_type.STOP or nod.edge == 1:
+            return
+        if mot.len == 0:
+            zero[0] += 1.0
+            return
+
+        if nod.strand == 1:
+            start = nod.ndx
+        else:
+            start = seq.slen-1-nod.ndx
+
+        # Stage 0:  Count all motifs.
+        # If a motif is detected, it is counted for every distance in stage 0.
+        # This is done to make sure off-distance good motifs are recognized.
+        if stage == 0:
+            for i in reversed(range(4)):
+                for j in range(start - 18 - i, start - 5 - i):
+                    if j < 0:
+                        continue
+                    if j <= start-16-i:
+                        spacendx = 3
+                    elif j <= start-14-i:
+                        spacendx = 2
+                    elif j >= start-7-i:
+                        spacendx = 1
+                    else:
+                        spacendx = 0
+                    for k in range(4):
+                        mcnt[i][k][seq._mer_ndx(j, length=i+3, strand=nod.strand)] += 1.0
+        # Stage 1:  Count only the best motif, but also count all its sub-motifs.
+        elif stage == 1:
+            mcnt[mot.len-3][mot.spacendx][mot.ndx] += 1.0;
+            for i in range(mot.len - 3):
+                for j in range(start-mot.spacer-mot.len, start-mot.spacer-i-2):
+                    if j < 0:
+                        continue
+                    if j <= start-16-i:
+                        spacendx = 3
+                    elif j <= start-14-i:
+                        spacendx = 2
+                    elif j >= start-7-i:
+                        spacendx = 1
+                    else:
+                        spacendx = 0
+                    mcnt[i][spacendx][seq._mer_ndx(j, length=i+3, strand=nod.strand)] += 1.0
+        # Stage 2:  Only count the highest scoring motif.
+        elif stage == 2:
+            mcnt[mot.len-3][mot.spacendx][mot.ndx] += 1.0
+
     cdef void _calc_dicodon_gene(self, Sequence seq, _node* nodes, int ipath) nogil:
         """Compute the dicodon frequency in genes and in the background.
 
@@ -2343,6 +2401,467 @@ cdef class TrainingInfo:
                 self.tinf.gene_dc[i] = 5.0
             elif self.tinf.gene_dc[i] < -5.0:
                 self.tinf.gene_dc[i] = -5.0
+
+    cdef void _count_upstream_composition(self, Sequence seq, int pos, int strand=1) nogil:
+        cdef int start
+        cdef int j
+        cdef int i     = 0
+
+        if strand == 1:
+            for j in range(1, 3):
+                if pos - j >= 0:
+                    self.tinf.ups_comp[i][seq.digits[pos-j] & 0b11] += 1
+                i += 1
+            for j in range(15, 45):
+                if pos - j >= 0:
+                    self.tinf.ups_comp[i][seq.digits[pos-j] & 0b11] += 1
+                i += 1
+        else:
+            start = seq.slen - 1 - pos
+            for j in range(1, 3):
+                if pos + j < seq.slen:
[…]
