### 2022-03-11 5a7c367
Add test to make sure `find_genes` works with a larger-than-default minimum gene size
