### 2022-03-11 9a2cc19
Port `record_overlapping_starts` and `tweak_final_starts` to allow customizing overlap
