+++ pyrodigal/_pyrodigal.pyx
@@ -64,7 +64,7 @@ from pyrodigal.prodigal cimport bitmap, dprog, gene, node, sequence
 from pyrodigal.prodigal.bitmap cimport bitmap_t
 from pyrodigal.prodigal.gene cimport _gene
 from pyrodigal.prodigal.metagenomic cimport NUM_META, _metagenomic_bin, initialize_metagenomic_bins
-from pyrodigal.prodigal.node cimport _motif, _node, MIN_EDGE_GENE, MIN_GENE, cross_mask
+from pyrodigal.prodigal.node cimport _motif, _node, MIN_EDGE_GENE, MIN_GENE, MAX_SAM_OVLP, cross_mask
 from pyrodigal.prodigal.sequence cimport _mask, node_type, rcom_seq
 from pyrodigal.prodigal.training cimport _training
 from pyrodigal._unicode cimport *
@@ -1705,6 +1705,58 @@ cdef class Nodes:
 
         return 0
 
+    cdef void _record_overlapping_starts(
+        self,
+        _training* tinf,
+        int flag,
+        int max_sam_overlap=MAX_SAM_OVLP
+    ) nogil:
+        cdef int    i
+        cdef int    j
+        cdef double sc
+        cdef double max_sc
+        cdef int    nn     = <int> self.length
+
+        for i in range(nn):
+            for j in range(3):
+                self.nodes[i].star_ptr[j] = -1
+            if self.nodes[i].type != node_type.STOP or self.nodes[i].edge == 1:
+                continue
+            if self.nodes[i].strand == 1:
+                max_sc = -100
+                for j in reversed(range(i+4)):
+                    if j >= nn or self.nodes[j].ndx > self.nodes[i].ndx+2:
+                        continue
+                    if self.nodes[j].ndx + max_sam_overlap < self.nodes[i].ndx:
+                        break
+                    if self.nodes[j].strand == 1 and self.nodes[j].type != node_type.STOP:
+                        if self.nodes[j].stop_val <= self.nodes[i].ndx:
+                            continue
+                        if flag == 0 and self.nodes[i].star_ptr[self.nodes[j].ndx%3] == -1:
+                            self.nodes[i].star_ptr[self.nodes[j].ndx%3] = j
+                        elif flag == 1:
+                            sc = self.nodes[j].cscore + self.nodes[j].sscore + node.intergenic_mod(&self.nodes[i], &self.nodes[j], tinf)
+                            if sc > max_sc:
+                                self.nodes[i].star_ptr[self.nodes[j].ndx%3] = j
+                                max_sc = sc
+            else:
+                max_sc = -100
+                for j in range(i-3, nn):
+                    if j < 0 or self.nodes[j].ndx < self.nodes[i].ndx-2:
+                        continue
+                    if self.nodes[j].ndx - max_sam_overlap > self.nodes[i].ndx:
+                        break
+                    if self.nodes[j].strand == -1 and self.nodes[j].type != node_type.STOP:
+                        if self.nodes[j].stop_val >= self.nodes[i].ndx:
+                            continue
+                        if flag == 0 and self.nodes[i].star_ptr[self.nodes[j].ndx%3] == -1:
+                            self.nodes[i].star_ptr[self.nodes[j].ndx%3] = j
+                        elif flag == 1:
+                            sc = self.nodes[j].cscore + self.nodes[j].sscore + node.intergenic_mod(&self.nodes[j], &self.nodes[i], tinf)
+                            if sc > max_sc:
+                                self.nodes[i].star_ptr[self.nodes[j].ndx%3] = j
+                                max_sc = sc
+
     cdef int _score(
[…]
