### 2022-03-11 9f98069
Add tests to make sure different `min_gene` and `max_overlap` values work as expected
