### 2022-03-11 17a1aec
Add `Predictions.write_scores` to write the start scores in tabular format like Prodigal
