+++ pyrodigal/_pyrodigal.pyx
@@ -94,6 +94,8 @@ import warnings
 import textwrap
 import threading
 
+from ._version import __version__
+
 # --- Module-level constants -------------------------------------------------
 
 cdef int    IDEAL_SINGLE_GENOME = 100000
@@ -3128,6 +3130,27 @@ METAGENOMIC_BINS = _m
 
 # --- Predictions ------------------------------------------------------------
 
+cdef list _RBS_MOTIF = [
+    "None", "GGA/GAG/AGG", "3Base/5BMM", "4Base/6BMM", "AGxAG", "AGxAG",
+    "GGA/GAG/AGG", "GGxGG", "GGxGG", "AGxAG", "AGGAG(G)/GGAGG",
+    "AGGA/GGAG/GAGG", "AGGA/GGAG/GAGG", "GGA/GAG/AGG", "GGxGG",
+    "AGGA", "GGAG/GAGG", "AGxAGG/AGGxGG", "AGxAGG/AGGxGG",
+    "AGxAGG/AGGxGG", "AGGAG/GGAGG", "AGGAG", "AGGAG", "GGAGG",
+    "GGAGG", "AGGAGG", "AGGAGG", "AGGAGG",
+]
+
+cdef list _RBS_SPACER = [
+    "None", "3-4bp", "13-15bp", "13-15bp", "11-12bp", "3-4bp",
+    "11-12bp", "11-12bp", "3-4bp", "5-10bp", "13-15bp", "3-4bp",
+    "11-12bp", "5-10bp", "5-10bp", "5-10bp", "5-10bp", "11-12bp",
+    "3-4bp", "5-10bp", "11-12bp", "3-4bp", "5-10bp", "3-4bp",
+    "5-10bp", "11-12bp", "3-4bp", "5-10bp",
+]
+
+cdef list _NODE_TYPE = [
+    "ATG", "GTG", "TTG" , "Edge",
+]
+
 cdef class Prediction:
     """A single predicted gene found by Prodigal.
     """
@@ -3187,7 +3210,7 @@ cdef class Prediction:
         """
         node = self.owner.nodes.nodes[self.gene.gene.start_ndx]
         start_type = 3 if node.edge else node.type
-        return ["ATG", "GTG", "TTG" , "Edge"][start_type]
+        return _NODE_TYPE[start_type]
 
     @property
     def rbs_motif(self):
@@ -3452,7 +3475,7 @@ cdef class Prediction:
         return protein
 
 cdef class Predictions:
-    """A list of predictions found by Prodigal in a single sequence.
+    """A list of predictions made by Prodigal on a single sequence.
 
     Attributes:
         sequence (`pyrodigal.Sequence`): The compressed input sequence for
@@ -3489,7 +3512,7 @@ cdef class Predictions:
         pred.gene = self.genes[index]
         return pred
 
-    cpdef ssize_t write_gff(self, object file, str prefix="gene_", str tool="pyrodigal") except -1:
+    cpdef ssize_t write_gff(self, object file, str prefix="gene_") except -1:
         """write_gff(self, file, prefix="gene_", width=60)\n--
 
         Write the predictions to ``file`` in General Feature Format.
@@ -3499,9 +3522,6 @@ cdef class Predictions:
                the features.
            prefix (`str`): The prefix to use to make identifiers for each
                predicted gene.
-           tool (`str`): The name of the gene detection tool, typically
-              ``pyrodigal``, but you can change to a different name if
-              you are building a larger pipeline.
 
         Returns:
             `int`: The number of bytes written to the file.
@@ -3515,7 +3535,8 @@ cdef class Predictions:
             n += file.write(prefix)
             n += file.write(str(i+1))
             n += file.write("\t")
-            n += file.write(tool)
[…]
