+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pyrodigal
-version = attr: pyrodigal.__version__
+version = attr: pyrodigal._version.__version__
 author = Martin Larralde
 author_email = martin.larralde@embl.de
 url = https://github.com/althonos/pyrodigal
