### 2022-03-11 709c81b
Fix `Predictions.write_scores` not sorting nodes by STOP position
