+++ pyrodigal/_pyrodigal.pyx
@@ -64,7 +64,7 @@ from pyrodigal.prodigal cimport bitmap, dprog, gene, node, sequence
 from pyrodigal.prodigal.bitmap cimport bitmap_t
 from pyrodigal.prodigal.gene cimport _gene
 from pyrodigal.prodigal.metagenomic cimport NUM_META, _metagenomic_bin, initialize_metagenomic_bins
-from pyrodigal.prodigal.node cimport _motif, _node, MIN_EDGE_GENE, MIN_GENE, MAX_SAM_OVLP, cross_mask
+from pyrodigal.prodigal.node cimport _motif, _node, MIN_EDGE_GENE, MIN_GENE, MAX_SAM_OVLP, cross_mask, compare_nodes, stopcmp_nodes
 from pyrodigal.prodigal.sequence cimport _mask, node_type, rcom_seq
 from pyrodigal.prodigal.training cimport _training
 from pyrodigal._unicode cimport *
@@ -1920,7 +1920,7 @@ cdef class Nodes:
     cdef int _sort(self) nogil except 1:
         """Sort all nodes in the vector by their index and strand.
         """
-        qsort(self.nodes, self.length, sizeof(_node), node.compare_nodes)
+        qsort(self.nodes, self.length, sizeof(_node), compare_nodes)
         return 0
 
     cdef int _reset_scores(self) nogil except 1:
@@ -3672,73 +3672,80 @@ cdef class Predictions:
         cdef double    rbs1
         cdef double    rbs2
 
-        if header:
-            # FIXME: missing some header data
-            # Sequence Data: seqnum=1;seqlen=20000;seqhdr="KK037166.1 Kutzneria sp. 744 genomic scaffold supercont1.1, whole genome shotgun sequence"
-            # Run Data: version=Prodigal.v2.6.3;run_type=Single;model="Ab initio";gc_cont=66.03;transl_table=11;uses_sd=0
-            n += file.write(f"# Sequence Data: seqlen={len(self.sequence)}\n")
-            n += file.write(f"# Run Data: version=pyrodigal.v{__version__};gc_cont={tinf.gc*100:.2f};transl_table={tinf.trans_table};uses_sd={tinf.uses_sd}\n")
-            # write column names
-            n += file.write(
-                "Beg\t" "End\t" "Std\t" "Total\t" "CodPot\t" "StrtSc\t"
-                "Codon\t" "RBSMot\t" "Spacer\t" "RBSScr\t" "UpsScr\t"
-                "TypeScr\t" "GCCont\n"
-            )
+        try:
+            # Sort and groupd nodes by STOP codon position
+            qsort(self.nodes.nodes, self.nodes.length, sizeof(_node), stopcmp_nodes)
+            # Write headers if requested
+            if header:
+                # FIXME: missing some header data
+                # Sequence Data: seqnum=1;seqlen=20000;seqhdr="KK037166.1 Kutzneria sp. 744 genomic scaffold supercont1.1, whole genome shotgun sequence"
+                # Run Data: version=Prodigal.v2.6.3;run_type=Single;model="Ab initio";gc_cont=66.03;transl_table=11;uses_sd=0
+                n += file.write(f"# Sequence Data: seqlen={len(self.sequence)}\n")
+                n += file.write(f"# Run Data: version=pyrodigal.v{__version__};gc_cont={tinf.gc*100:.2f};transl_table={tinf.trans_table};uses_sd={tinf.uses_sd}\n")
+                # write column names
+                n += file.write(
+                    "Beg\t" "End\t" "Std\t" "Total\t" "CodPot\t" "StrtSc\t"
+                    "Codon\t" "RBSMot\t" "Spacer\t" "RBSScr\t" "UpsScr\t"
[…]
