### 2022-03-11 44e8186
Remove cache in GitHub Actions to fix outdated built files crashing `setup.py`
