### 2022-03-11 649f75b
Patch version in `setup.cfg` to fix CI temporarily
