+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pyrodigal
-version = attr: pyrodigal._version.__version__
+version = 0.7.0 #attr: pyrodigal._version.__version__
 author = Martin Larralde
 author_email = martin.larralde@embl.de
 url = https://github.com/althonos/pyrodigal
