### 2022-03-11 80b34bf
Skip testing `Predictions.write_scores` on PyPy because of numerical precision issues
