+++ pyrodigal/tests/test_genes.py
@@ -57,6 +57,7 @@ class TestGenes(unittest.TestCase):
         if sys.version_info >= (3, 6):
             self.assertIsInstance(self.genes, collections.abc.Reversible)
 
+    @unittest.skipIf(sys.implementation.name != 'cpython', 'numerical precision issues on PyPy')
     def test_write_scores(self):
 
         buffer = io.StringIO()
