+++ .github/workflows/test.yml
@@ -49,12 +49,12 @@ jobs:
         path: ~/.cache/pip
         key: ${{ runner.os }}-pip-${{ matrix.python-version }}
         restore-keys: ${{ runner.os }}-pip-${{ matrix.python-version }}
-    # - name: Cache built C files
-    #   uses: actions/cache@v2
-    #   with:
-    #     path: build
-    #     key: ${{ runner.os }}-build-${{ matrix.python-version }}
-    #     restore-keys: ${{ runner.os }}-build-${{ matrix.python-version }}
+    - name: Cache built C files
+      uses: actions/cache@v2
+      with:
+        path: build
+        key: ${{ runner.os }}-build-${{ matrix.python-version }}-${{ secrets.CACHE_VERSION }}
+        restore-keys: ${{ runner.os }}-build-${{ matrix.python-version }}-${{ secrets.CACHE_VERSION }}
     - name: Setup Python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
       with:
@@ -115,12 +115,12 @@ jobs:
         path: ~/.cache/pip
         key: ${{ runner.os }}-pip-${{ matrix.python-version }}
         restore-keys: ${{ runner.os }}-pip-${{ matrix.python-version }}
-    # - name: Cache built C files
-    #   uses: actions/cache@v2
-    #   with:
-    #     path: build
-    #     key: ${{ runner.os }}-build-${{ matrix.python-version }}
-    #     restore-keys: ${{ runner.os }}-build-${{ matrix.python-version }}
+    - name: Cache built C files
+      uses: actions/cache@v2
+      with:
+        path: build
+        key: ${{ runner.os }}-build-${{ matrix.python-version }}-${{ secrets.CACHE_VERSION }}
+        restore-keys: ${{ runner.os }}-build-${{ matrix.python-version }}-${{ secrets.CACHE_VERSION }}
     - name: Setup Python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
       with:
@@ -174,12 +174,12 @@ jobs:
       uses: actions/checkout@v2
       with:
         submodules: true
-    # - name: Cache built C files
-    #   uses: actions/cache@v2
-    #   with:
-    #     path: build
-    #     key: ${{ runner.os }}-build-${{ matrix.python-version }}
-    #     restore-keys: ${{ runner.os }}-build-${{ matrix.python-version }}
+    - name: Cache built C files
+      uses: actions/cache@v2
+      with:
+        path: build
+        key: ${{ runner.os }}-build-${{ matrix.python-version }}-${{ secrets.CACHE_VERSION }}
+        restore-keys: ${{ runner.os }}-build-${{ matrix.python-version }}-${{ secrets.CACHE_VERSION }}
     - name: Setup Python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
       with:
