### 2022-03-11 52a3246
Attempt fixing circular import by including version tag directly in `_pyrodigal.pyx`
