+++ .github/workflows/test.yml
@@ -38,6 +38,12 @@ jobs:
         - python-version: pypy-3.7
           python-release: v3.7
           python-impl: PyPy
+        - python-version: pypy-3.8
+          python-release: v3.8
+          python-impl: PyPy
+        - python-version: pypy-3.9
+          python-release: v3.9
+          python-impl: PyPy
     steps:
     - name: Checkout code
       uses: actions/checkout@v2
@@ -104,6 +110,12 @@ jobs:
         - python-version: pypy-3.7
           python-release: v3.7
           python-impl: PyPy
+        - python-version: pypy-3.8
+          python-release: v3.8
+          python-impl: PyPy
+        - python-version: pypy-3.9
+          python-release: v3.9
+          python-impl: PyPy
     steps:
     - name: Checkout code
       uses: actions/checkout@v2
@@ -168,6 +180,12 @@ jobs:
           python-impl: CPython
         - python-version: pypy-3.7
           python-release: v3.7
+          python-impl: PyPy
+        - python-version: pypy-3.8
+          python-release: v3.8
+          python-impl: PyPy
+        - python-version: pypy-3.9
+          python-release: v3.9
           python-impl: PyPy
     steps:
     - name: Checkout code
