### 2022-03-11 ed009d8
Merge `package.yml` and `changelog.yml` Actions workflows
