### 2022-03-12 12b6b5f
Add `Sequence.shine_dalgarno` method with unit tests prior to refactoring
