### 2022-03-12 ba1d5a4
Mark some `ConnectionScorer` functions as raising no exception
