### 2022-03-12 acd9b76
Add `__repr__` methods for some types missing one
