### 2022-03-12 61e84d6
Rewrite `Prediction.rbs_motif` and `Prediction.rbs_spacer` to avoid reading string data
