### 2022-03-12 4b4c063
Merge `Genes` and `Predictions` classes to make public API less complicated
