+++ pyrodigal/_pyrodigal.pyx
@@ -2182,14 +2182,26 @@ cdef class Gene:
 
     def __repr__(self):
         ty = type(self)
-        return "<{}.{} begin={!r} end={!r}>".format(
+        return "<{}.{} begin={!r} end={!r} strand={:+} start_type={!r} rbs_motif={!r} rbs_spacer={!r}>".format(
             ty.__module__,
             ty.__name__,
             self.begin,
             self.end,
+            self.strand,
+            self.start_type,
+            self.rbs_motif,
+            self.rbs_spacer,
         )
 
     # --- Properties ---------------------------------------------------------
+
+    @property
+    def _gene_data(self):
+        return self.gene.gene_data.decode('ascii')
+
+    @property
+    def _score_data(self):
+        return self.gene.score_data.decode('ascii')
 
     @property
     def begin(self):
@@ -2204,21 +2216,357 @@ cdef class Gene:
         return self.gene.end
 
     @property
-    def start_ndx(self):
-        """`int`: The index of the start node in the `Nodes` list.
+    def strand(self):
+        """`int`: *-1* if the gene is on the reverse strand, *+1* otherwise.
         """
-        return self.gene.start_ndx
+        return self.owner.nodes.nodes[self.gene.start_ndx].strand
+
+    @property
+    def partial_begin(self):
+        """`bool`: whether the gene overlaps with the start of the sequence.
+        """
+        if self.strand == 1:
+            return self.owner.nodes.nodes[self.gene.start_ndx].edge == 1
+        else:
+            return self.owner.nodes.nodes[self.gene.stop_ndx].edge == 1
+
+    @property
+    def partial_end(self):
+        """`bool`: whether the gene overlaps with the end of the sequence.
+        """
+        if self.strand == 1:
+            return self.owner.nodes.nodes[self.gene.stop_ndx].edge == 1
+        else:
+            return self.owner.nodes.nodes[self.gene.start_ndx].edge == 1
+
+    @property
+    def start_type(self):
+        """`str`: The start codon of this gene.
+
+        Can be one of ``ATG``, ``GTG`` or ``TTG``, or ``Edge`` if the
+        `OrfFinder` has been initialized in open ends mode and the gene
+        starts right at the beginning of the input sequence.
+
+        """
+        node = self.owner.nodes.nodes[self.gene.start_ndx]
+        start_type = 3 if node.edge else node.type
+        return _NODE_TYPE[start_type]
+
+    @property
+    def rbs_motif(self):
+        """``str``, optional: The motif of the Ribosome Binding Site.
+
+        Possible non-`None` values are ``GGA/GAG/AGG``, ``3Base/5BMM``,
+        ``4Base/6BMM``, ``AGxAG``, ``GGxGG``, ``AGGAG(G)/GGAGG``, ``AGGA``,
+        ``AGGA/GGAG/GAGG``, ``GGAG/GAGG``, ``AGGAG/GGAGG``, ``AGGAG``,
+        ``GGAGG`` or ``AGGAGG``.
+
+        """
+        assert self.owner is not None
+        assert self.owner.nodes is not None
+        assert self.owner.nodes.nodes != NULL
+
+        cdef char[10]   qt
+        cdef _node*     node = &self.owner.nodes.nodes[self.gene.start_ndx]
+        cdef _training* tinf = self.owner.training_info.tinf
+        cdef double     rbs1 = tinf.rbs_wt[node.rbs[0]] * tinf.st_wt
+        cdef double     rbs2 = tinf.rbs_wt[node.rbs[1]] * tinf.st_wt
+
+        if tinf.uses_sd:
[…]
