### 2022-03-12 5759136
Make `Gene` constructor require `Sequence`, `TrainingInfo` and `Nodes` objects
