### 2022-03-12 3f3eaac
Turn `add_genes` into a method of `Genes` objects
