### 2022-03-12 aa01348
Build `Gene._score_data` and `Gene._gene_data` on demand instead of allocating strings for every gene
