### 2022-03-12 fb22da0
Use a more compact data structure to store gene data in memory
