### 2022-03-12 c9478c3
Disable fail-fast in `test.yml` Actions workflow to simplify debugging
