### 2022-03-12 d7568f3
Fix `Nodes._calc_orf_gc` causing sequence memory to be read after buffer end
