### 2022-03-12 35b6369
Remove `Predictions` from the Sphinx documentation
