+++ docs/api/predictions.rst
@@ -1,13 +1,0 @@
-Predictions
-===========
-
-.. currentmodule:: pyrodigal
-
-
-.. autoclass:: pyrodigal.Predictions
-   :special-members: __init__
-   :members:
-
-.. autoclass:: pyrodigal.Prediction
-   :special-members: __init__
-   :members:
