### 2022-03-12 86819cb
Remove `Pyrodigal` class in preparation for `v0.7.0`
