### 2022-03-12 e43e2c9
Fix rendering of some docstrings in `pyrodigal.Genes`
