### 2022-03-12 8bc8fae
Implement `Sequence.__str__` to get a string back from a sequence
