### 2022-03-12 bc10f6d
Update type hints in `pyrodigal/_pyrodigal.pyi` with newest version of API
