### 2022-03-12 a63fe5a
Add option to write the scores file in the CLI
