+++ pyrodigal/tests/fasta.py
@@ -17,13 +17,16 @@ def parse(path):
         id_ = None
         seq = []
         for line in file:
+            l = line.strip()
             if line.startswith(">"):
                 if id_ is not None:
                     yield Record(id_, "".join(seq), desc)
                 id_ = line[1:].split()[0].strip()
                 desc = " ".join(line[1:].split(maxsplit=1))
                 seq = []
-            else:
-                seq.append(line.strip())
+            elif l:
+                seq.append(l)
         if id_ is not None:
             yield Record(id_, "".join(seq), desc)
+        elif seq:
+            raise ValueError("not in FASTA format")
