+++ .github/workflows/package.yml
@@ -78,9 +78,6 @@ jobs:
         - 3.8
         - 3.9
         - '3.10'
-        - pypy-3.7
-        - pypy-3.8
-        - pypy-3.9
     steps:
     - name: Setup Python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
@@ -91,6 +88,8 @@ jobs:
       with:
           name: dist
           path: dist
+    - name: Update pip to latest version
+      run: python -m pip install -U pip setuptools wheel
     - name: Install built wheel
       run: python -m pip install --no-index --find-links=dist pyrodigal
     - name: Run tests without coverage
