### 2022-03-12 72105a4
Skip `TestGenes.test_write_scores` on Aarch64
