### 2022-03-13 9a2eef9
Update `figure2.svg` to use latest benchmark results
