+++ benches/connection_scoring/plot.py
@@ -58,13 +58,9 @@ for color, (backend, group) in zip(
     Y = numpy.array([r["mean"] for r in group])
     reg = scipy.stats.linregress(X, Y)
     plt.plot([ 0, max(X) ], [ reg.intercept, reg.slope*max(X) + reg.intercept ], color=color, linestyle="--", marker="")
-<<<<<<< HEAD
     # ci = [1.96 * r["stddev"] / math.sqrt(len(r["times"])) for r in group]
     plt.scatter(X, Y, marker="+", color=color, label=f"{backend} (R²={reg.rvalue**2:.3f})")
-=======
-    ci = [1.96 * r["stddev"] / math.sqrt(len(r["times"])) for r in group]
-    plt.errorbar(X, Y, ci, linestyle='', marker="+", color=color, elinewidth=0.3, ecolor='black', label=f"{backend} ($R^2$={reg.rvalue**2:.3f})")
->>>>>>> d7a9c03 (Update connection scoring benchmark plot to show R² instead of R)
+
 plt.legend()
 plt.xlabel("Nucleotide count (Mbp)")
 plt.ylabel("Time (s)")
