### 2022-03-13 a2daf0d
Update `pkg.yml` Actions workflow to update the `python-pyrodigal` AUR package
