### 2022-03-13 64be473
Add badge listing AUR package version on `README.md` [ci skip]
