### 2022-03-14 1906f10
Fix `TestGenes.test_write_scores` to skip comparing formatted float values
