### 2022-03-14 59010c3
Fix bug causing `Nodes.copy` to fail when copying an empty collection
