### 2022-03-14 ad36a0c
Inline functions used for nucleotide matching using an additional C header
