+++ pyrodigal/_pyrodigal.pyx
@@ -67,6 +67,19 @@ from pyrodigal.prodigal.node cimport _motif, _node, MIN_EDGE_GENE, MIN_GENE, MAX
 from pyrodigal.prodigal.sequence cimport _mask, node_type, rcom_seq
 from pyrodigal.prodigal.training cimport _training
 from pyrodigal._unicode cimport *
+from pyrodigal._sequence cimport (
+    nucleotide,
+    _is_a,
+    _is_g,
+    _is_gc,
+    _is_stop,
+    _is_start,
+    _is_atg,
+    _is_ttg,
+    _is_gtg,
+    _letters,
+    _complement
+)
 
 IF TARGET_CPU == "x86":
     from pyrodigal.cpu_features.x86 cimport GetX86Info, X86Info
@@ -214,22 +227,6 @@ cdef enum:
     C = 0b010
     T = 0b011
     N = 0b110  # use 6 so that N & 0x3 == C
-
-cdef uint8_t _complement[N+1]
-for i in range(N+1):
-    _complement[i] = N
-_complement[A] = T
-_complement[T] = A
-_complement[C] = G
-_complement[G] = C
-
-cdef Py_UCS4 _letters[N+1]
-for i in range(N+1):
-    _letters[i] = "N"
-_letters[A] = "A"
-_letters[T] = "T"
-_letters[C] = "C"
-_letters[G] = "G"
 
 cdef class Sequence:
     """A digitized input sequence.
@@ -422,149 +419,7 @@ cdef class Sequence:
             memset(self.digits, 0, slen * sizeof(uint8_t))
         return 0
 
-    cdef inline bint _is_a(self, int i, int strand = 1) nogil:
-        cdef uint8_t x
-
-        if strand == 1:
-            return self.digits[i] == A
-        else:
-            return self.digits[self.slen - 1 - i] == T
-
-    cdef inline bint _is_g(self, int i, int strand = 1) nogil:
-        cdef uint8_t x
-
-        if strand == 1:
-            return self.digits[i] == G
-        else:
-            return self.digits[self.slen - 1 - i] == C
-
-    cdef inline bint _is_gc(self, int i, int strand = 1) nogil:
-        cdef uint8_t x
-
-        if strand == 1:
-            x = self.digits[i]
-        else:
-            x = self.digits[self.slen - 1 - i]
-
-        # NB(@althonos): In the original Prodigal implementation, any unknown
-        #                character gets encoded as a C, so it gets counted
-        #                when computing the GC percent. We reproduce this
-        #                behaviour here, but a better solution would be to
-        #                count only known letters.
-        return x == C or x == G or x == N
-
-    cdef inline bint _is_start(self, int i, int tt, int strand = 1) nogil:
-        cdef uint8_t x0
-        cdef uint8_t x1
-        cdef uint8_t x2
-
-        if strand == 1:
-            x0 = self.digits[i]
-            x1 = self.digits[i+1]
-            x2 = self.digits[i+2]
-        else:
-            x0 = _complement[self.digits[self.slen - 1 - i]]
-            x1 = _complement[self.digits[self.slen - 2 - i]]
-            x2 = _complement[self.digits[self.slen - 3 - i]]
-
-        # ATG
-        if x0 == A and x1 == T and x2 == G:
-            return True
-        # Codes that only use ATG
-        if tt == 6 or tt == 10 or tt == 14 or tt == 15 or tt == 16 or tt == 2:
-            return False
-        # GTG
-        if x0 == G and x1 == T and x2 == G:
-            return not (tt == 1 or tt == 3 or tt == 12 or tt == 2)
-        # TTG
-        if x0 == T and x1 == T and x2 == G:
[…]
