### 2022-03-14 236263c
Fix debug flags used to build extension with `--debug` argument
