+++ setup.cfg
@@ -60,6 +60,7 @@ plugins = Cython.Coverage
 
 [coverage:report]
 include = pyrodigal/*
+omit = pyrodigal/tests/*
 show_missing = true
 exclude_lines =
     pragma: no cover
