### 2022-03-14 1f9b509
Fix coverage not being properly collected in GitHub Actions
