### 2022-03-14 1287df6
Implement `_mer_ndx` in `_sequence.h` so that it can be inlined
