### 2022-03-14 d81ce87
Use bit-twiddling instead of LUT to complement nucleotides when possible in `_sequence.h`
