### 2022-03-14 36d9823
Use `nucleotide` enum values defined in `_sequence.h`
