### 2022-03-14 de16af2
Skip measuring coverage for PyPy
