### 2022-03-14 845d9fe
Only run coverage measurement on the `main` branch
