+++ .github/workflows/test.yml
@@ -206,6 +206,7 @@ jobs:
     name: Coverage (Linux)
     runs-on: ubuntu-latest
     needs: test_linux
+    if: "github.ref == 'refs/heads/main'"
     env:
       OS: Linux
     strategy:
