+++ .github/workflows/package.yml
@@ -359,7 +359,7 @@ jobs:
           name: dist
           path: dist
     - name: Publish distributions to PyPI
-      if: startsWith(github.ref, 'refs/tags/v')
+      if: "startsWith(github.ref, 'refs/tags/v')"
       uses: pypa/gh-action-pypi-publish@master
       with:
         user: __token__
@@ -403,7 +403,7 @@ jobs:
           path: dist
     - name: Extract package version
       run: echo PKGVER=$(python setup.py --version) >> $GITHUB_ENV
-    - name: Compute SHA256 of source distribution 
+    - name: Compute SHA256 of source distribution
       run: echo SHA256=$(sha256sum dist/pyrodigal-${{ env.PKGVER }}.tar.gz | cut -f1 -d' ') >> $GITHUB_ENV
     - name: Generate PKGBUILD
       run: sed -e "s/%pkgver/${{ env.PKGVER }}/g" -e "s/%sha256sum/${{ env.SHA256 }}/g" pkg/aur/PKGBUILD.in > pkg/aur/PKGBUILD
