### 2022-03-14 b068dd9
Fix `Gene.translate` ignoring the `unknown_residue` argument value
