### 2022-03-14 8e1f29c
Make Cython compile with additional warnings when in debug mode
