+++ .github/workflows/test.yml
@@ -70,8 +70,8 @@ jobs:
       run: python -m pip install -U pip wheel setuptools
     - name: Install Python requirements
       run: python -m pip install -r .github/workflows/requirements.txt
-    - name: Build C extension
-      run: python setup.py build_ext --inplace
+    - name: Build C extension in debug mode
+      run: python setup.py build_ext --inplace --debug
     - name: Test without coverage
       run: python -m unittest discover -vv
 
@@ -136,8 +136,8 @@ jobs:
       run: python -m pip install -U pip wheel setuptools
     - name: Install Python requirements
       run: python -m pip install -r .github/workflows/requirements.txt
-    - name: Build C extension
-      run: python setup.py build_ext --inplace
+    - name: Build C extension in debug mode
+      run: python setup.py build_ext --inplace --debug
     - name: Test without coverage
       run: python -m unittest discover -vv
 
@@ -197,8 +197,8 @@ jobs:
       run: python -m pip install -U pip wheel setuptools
     - name: Install Python requirements
       run: python -m pip install -r .github/workflows/requirements.txt
-    - name: Build C extension
-      run: python setup.py build_ext --inplace
+    - name: Build C extension in debug mode
+      run: python setup.py build_ext --inplace --debug
     - name: Test without coverage
       run: python -m unittest discover -vv
 
