### 2022-03-14 d72b130
Build extension in release mode on Windows Actions runner
