### 2022-03-14 a284c96
Fix memory leak caused by memory not being freed in `calc_most_gc_frame`
