### 2022-03-14 14e87a3
Fix older compilers failing to use `inline` functions in `_sequence.h`
