### 2022-03-14 4c4b9c3
Update `PKGBUILD.in` to run all tests during `check` phase [ci skip]
