+++ pkg/aur/PKGBUILD.in
@@ -24,7 +24,7 @@ check() {
     local pyver=$(python -c 'import sys; print("{}.{}".format(*sys.version_info[:2]))')
     local machine=$(python -c 'import platform; print(platform.machine())')
     cd "${srcdir}/${_name}-${pkgver}/build/lib.linux-${machine}-${pyver}"
-    python -m unittest ${_name}.tests.test_orf_finder
+    python -m unittest ${_name}.tests
 }
 
 package() {
