### 2022-03-14 b54cfb8
Remove the `embedsignature` Cython directive from `_pyrodigal.pyx`
