### 2022-03-15 4ba781f
Reword `Memory` subsection of `README.md` [ci skip]
