### 2022-03-15 1c2f9b8
Use a generic implementation of the connection scoring filter for unsupported CPUs
