### 2022-03-15 9c17748
Add functions to compute the intergenic node modifier as `Node` static methods
