### 2022-03-15 34044ae
Fix some repeated calls to `_mer_ndx` in `TrainingInfo._update_motif_counts`
