### 2022-03-15 1d87ee6
Fix `_sequence.h` not being inlined as expected
