+++ pyrodigal/_pyrodigal.pyx
@@ -413,7 +413,14 @@ cdef class Sequence:
             memset(self.digits, 0, slen * sizeof(uint8_t))
         return 0
 
-    cdef char _amino(self, int i, int tt, int strand = 1, bint is_init = False, char unknown_residue = b"X") nogil:
+    cdef char _amino(
+        self,
+        int i,
+        int tt,
+        int strand = 1,
+        bint is_init = False,
+        char unknown_residue = b"X"
+    ) nogil:
         cdef uint8_t x0
         cdef uint8_t x1
         cdef uint8_t x2
@@ -501,7 +508,13 @@ cdef class Sequence:
 
         return unknown_residue
 
-    cdef int _shine_dalgarno_exact(self, int pos, int start, _training* tinf, int strand=1) nogil:
+    cdef int _shine_dalgarno_exact(
+        self,
+        const int pos,
+        const int start,
+        const _training* tinf,
+        const int strand
+    ) nogil:
         cdef int i
         cdef int j
         cdef int k
@@ -515,16 +528,12 @@ cdef class Sequence:
         cdef int cur_ctr
         cdef int dis_flag
 
-        limit = start - 4 - pos
-        if limit > 6:
-            limit = 6
-        for i in range(limit, 6):
-            match[i] = -10
+        # reset the match array
+        match[0] = match[1] = match[2] = match[3] = match[4] = match[5] = -10
 
         # Compare the 6-base region to AGGAGG
+        limit = min(6, start - 4 - pos)
         for i in range(limit):
-            if pos + i < 0:
-                continue
             if i%3 == 0 and _is_a(self.digits, self.slen, pos+i, strand):
                 match[i] = 2
             elif i%3 != 0 and _is_g(self.digits, self.slen, pos+i, strand):
@@ -620,7 +629,13 @@ cdef class Sequence:
 
         return max_val
 
-    cdef int _shine_dalgarno_mm(self, int pos, int start, _training* tinf, int strand=1) nogil:
+    cdef int _shine_dalgarno_mm(
+        self,
+        const int pos,
+        const int start,
+        const _training* tinf,
+        const int strand
+    ) nogil:
         cdef int i
         cdef int j
         cdef int k
@@ -634,20 +649,16 @@ cdef class Sequence:
         cdef int cur_ctr
         cdef int dis_flag
 
-        limit = start - 4 - pos
-        if limit > 6:
-            limit = 6
-        for i in range(limit, 6):
-            match[i] = -10
+        # reset the match array
+        match[0] = match[1] = match[2] = match[3] = match[4] = match[5] = -10
 
         # Compare the 6-base region to AGGAGG
+        limit = min(6, start - 4 - pos)
         for i in range(limit):
-            if pos + i < 0:
-                continue
             if i%3 == 0:
-                match[i] = 2 if _is_a(self.digits, self.slen, pos+i, strand) else -3
+                match[i] = -3 + 5*_is_a(self.digits, self.slen, pos+i, strand)
             else:
-                match[i] = 3 if _is_g(self.digits, self.slen, pos+i, strand) else -2
+                match[i] = -2 + 5*_is_g(self.digits, self.slen, pos+i, strand)
 
         # Find the maximally scoring motif
         max_val = 0
@@ -923,7 +934,14 @@ cdef class ConnectionScorer:
         with nogil:
             self._compute_skippable(min, i)
 
-    cdef int _score_connections(self, Nodes nodes, int min, int i, _training* tinf, bint final=False) nogil:
+    cdef int _score_connections(
+        self,
+        Nodes nodes,
+        const int min,
+        const int i,
+        const _training* tinf,
+        const bint final
+    ) nogil:
         cdef int j
 
[…]
