### 2022-03-15 d6af859
Rewrite `dprog.score_connection` as a `ConnectionScorer` static method
