+++ pyrodigal/_pyrodigal.pyx
@@ -810,6 +810,8 @@ cdef enum simd_backend:
 
 cdef class ConnectionScorer:
 
+    # --- Magic methods ------------------------------------------------------
+
     def __cinit__(self):
         self.capacity = 0
         self.skip_connection = self.skip_connection_raw = NULL
@@ -871,7 +873,9 @@ cdef class ConnectionScorer:
         PyMem_Free(self.node_frames_raw)
         PyMem_Free(self.skip_connection_raw)
 
-    cdef int _index(self, Nodes nodes) nogil except 1:
+    # --- C interface --------------------------------------------------------
+
+    cdef int _index(self, Nodes nodes) nogil except -1:
         cdef size_t i
         # reallocate if needed
         if self.capacity < nodes.length:
@@ -906,11 +910,11 @@ cdef class ConnectionScorer:
         # return 0 if no exceptions were raised
         return 0
 
-    def index(self, Nodes nodes not None):
-        with nogil:
-            self._index(nodes)
-
-    cdef int _compute_skippable(self, int min, int i) nogil:
+    cdef int _compute_skippable(
+        self,
+        int min,
+        int i
+    ) nogil:
         memset(&self.skip_connection[min], 0, sizeof(uint8_t) * (i - min))
         IF AVX2_BUILD_SUPPORT:
             if self.backend == simd_backend.AVX2:
@@ -927,12 +931,174 @@ cdef class ConnectionScorer:
         skippable_generic(self.node_strands, self.node_types, self.node_frames, min, i, self.skip_connection)
         return 0
 
-    def compute_skippable(self, int min, int i):
-        assert self.skip_connection != NULL
-        assert i < <int> self.capacity
-        assert min <= i
-        with nogil:
-            self._compute_skippable(min, i)
+    @staticmethod
+    cdef void _score_connection(
+        Nodes nodes,
+        const int p1,
+        const int p2,
+        const _training* tinf,
+        const bint final
+    ) nogil:
+
+        cdef _node* n1 = &nodes.nodes[p1]
+        cdef _node* n2 = &nodes.nodes[p2]
+        cdef _node* n3
+
+        cdef int i
+        cdef int bnd
+        cdef int ovlp  = 0
+        cdef int maxfr = -1
+        cdef int left  = n1.ndx
+        cdef int right = n2.ndx
+
+        cdef double maxval
+        cdef double score   = 0.0
+        cdef double scr_mod = 0.0
+
+        # NOTE(@althonos): We can skip checking for invalid connections here
+        #                  because the connection scorer already checked for
+        #                  them using the SIMD filter
+
+        # --- Edge Artifacts ---
+        if n1.traceb == -1 and n1.strand == 1 and n1.type == node_type.STOP:
+            return
+        elif n1.traceb == -1 and n1.strand == -1 and n1.type != node_type.STOP:
+            return
+
+        # --- Genes ---
+        # 5'fwd->3'fwd
+        elif n1.strand == 1 and n2.strand == 1 and n1.type != node_type.STOP and n2.type == node_type.STOP:
+            if n2.stop_val >= n1.ndx:
+                return
+            right += 2
+            if final:
+                score = n1.cscore + n1.sscore
+            else:
+                scr_mod = tinf.bias[0]*n1.gc_score[0] + tinf.bias[1]*n1.gc_score[1] + tinf.bias[2]*n1.gc_score[2]
+        # 3'rev->5'rev */
+        elif n1.strand == -1 and n2.strand == -1 and n1.type == node_type.STOP and n2.type != node_type.STOP:
+            if n1.stop_val <= n2.ndx:
+                return
[…]
