### 2022-03-15 f3f6e71
Add option to use original Prodigal code in `ConnectionScoring` for testing
