### 2022-03-15 f964079
Test that the generic connection filter produces the same score as the Prodigal code
