### 2022-03-15 cd5d3ef
Update `connection_scoring` benchmark to test the generic filter
