### 2022-03-15 5b1c6b7
Update `neon` SIMD code to use generic filter on edges
