### 2022-03-15 7cec8f6
Make `ConnectionScorer._index` a no-op when using Prodigal backend
