### 2022-03-15 9666df6
Replace `add_nodes` with `Nodes.extract` in `connection_scoring` benchmark code
