### 2022-03-15 172297d
Update plot script in `connection_scoring` benchmark
