### 2022-03-15 1327832
Use original Prodigal code instead of the generic filter when SIMD is unavailable
