### 2022-03-15 d6654d1
Disable coverage measurement since it takes too long
