### 2022-03-15 252f5db
Return double precision numbers from `Node._intergenic_mod` functions
