### 2022-03-15 c693dce
Add bench plot for next patch release
