+++ benches/connection_scoring/plot.py
@@ -43,7 +43,7 @@ for color, (backend, group) in zip(
     reg = scipy.stats.linregress(X, Y)
     plt.plot([ 0, max(X) ], [ reg.intercept, reg.slope*max(X) + reg.intercept ], color=color, linestyle="--", marker="")
     # ci = [1.96 * r["stddev"] / math.sqrt(len(r["times"])) for r in group]
-    plt.scatter(X, Y, marker="+", color=color, label=f"{backend} (R={reg.rvalue:.3f})")
+    plt.scatter(X, Y, marker="+", color=color, label=f"{backend} (R²={reg.rvalue**2:.3f})")
 plt.legend()
 plt.xlabel("Node count")
 plt.ylabel("Time (s)")
@@ -60,7 +60,7 @@ for color, (backend, group) in zip(
     reg = scipy.stats.linregress(X, Y)
     plt.plot([ 0, max(X) ], [ reg.intercept, reg.slope*max(X) + reg.intercept ], color=color, linestyle="--", marker="")
     # ci = [1.96 * r["stddev"] / math.sqrt(len(r["times"])) for r in group]
-    plt.scatter(X, Y, marker="+", color=color, label=f"{backend} (R={reg.rvalue:.3f})")
+    plt.scatter(X, Y, marker="+", color=color, label=f"{backend} (R²={reg.rvalue**2:.3f})")
 plt.legend()
 plt.xlabel("Nucleotide count (Mbp)")
 plt.ylabel("Time (s)")
