### 2022-04-04 #4296 @editorialbot
[REVIEW]: Pyrodigal: Python bindings and interface to Prodigal, an efficient method for gene prediction in prokaryotes.
**Submitting author:** @althonos (<a href="http://orcid.org/0000-0002-3947-4444">Martin Larralde</a>)
**Repository:** https://github.com/althonos/pyrodigal/
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
