### 2022-04-05 c883d99
Fix random behaviour caused by occasional out-of-bound access while in meta mode
