### 2022-04-05 1f1528e
Fix `OrfFinder.find_genes` not rebuilding node scores of the maximum phase in meta mode
