### 2022-04-05 069d11c
Add `Gene.score` property to get a gene score as reported in the score data
