### 2022-04-06 b05826b
Fix rescoring of best nodes in `OrfFinder._find_genes_meta`
