+++ pyrodigal/_pyrodigal.pyx
@@ -4394,12 +4394,10 @@ cdef class OrfFinder:
             min_edge_gene=self.min_edge_gene
         )
         nodes._sort()
+        # rescore nodes
+        scorer._index(nodes)
+        nodes._reset_scores()
         nodes._score(sequence, tinf, closed=self.closed, is_meta=True)
-        # rescore nodes (this is not done in the Prodigal code, but if 
-        # we want to provide accurate scores in the `Node` objects we
-        # have to do it)
-        nodes._record_overlapping_starts(tinf, True, self.max_overlap)
-        nodes._dynamic_programming(tinf, scorer, final=True)
         
         # return the max phase on success
         return max_phase
