### 2022-04-06 cab5bb4
Add regression test using data from #13
