### 2022-04-07 6013273
Implement pickle protocol for `pyrodigal.Nodes`
