### 2022-04-07 382787a
Implement pickle protocol for `pyrodigal.TrainingInfo`
