### 2022-04-08 069d959
Implement pickle protocol for `pyrodigal.OrfFinder`
