### 2022-04-08 01f1b84
Implement pickle protocol for `pyrodigal.Sequence`
