### 2022-04-08 a3befd2
Implement `TrainingInfo.__sizeof__` as a `cpdef` function
