### 2022-04-08 9dfebbd
Implement pickle protocol for `pyrodigal.Genes`
