### 2022-04-11 e6836a8
Avoid naive search for mask on every position when region-masking is enabled
