### 2022-04-11 96b6e2c
Add some magic methods to `Mask`
