+++ pyrodigal/tests/__init__.py
@@ -2,6 +2,7 @@ from . import (
     test_connection_scorer,
     test_gene,
     test_genes,
+    test_mask,
     test_nodes,
     test_orf_finder,
     test_sequence,
@@ -12,6 +13,7 @@ def load_tests(loader, suite, pattern):
     suite.addTests(loader.loadTestsFromModule(test_connection_scorer))
     suite.addTests(loader.loadTestsFromModule(test_gene))
     suite.addTests(loader.loadTestsFromModule(test_genes))
+    suite.addTests(loader.loadTestsFromModule(test_mask))
     suite.addTests(loader.loadTestsFromModule(test_nodes))
     suite.addTests(loader.loadTestsFromModule(test_orf_finder))
     suite.addTests(loader.loadTestsFromModule(test_sequence))
