### 2022-04-11 918e558
Expose `TRANSLATION_TABLES` and `METAGENOMIC_BINS` constants in `pyrodigal` module
