### 2022-04-11 5669389
Wrap exceptions in `pyrodigal.cli.main` to display a better error message
