### 2022-04-11 5e5e9d5
Add missing `pyrodigal.tests.test_mask` file
