### 2022-04-13 9033027
Implement buffer protocol for `Sequence` and refactor allocation in `Nodes` and `Genes`
