### 2022-04-13 7ef7973
Refactor common code between `Sequence.from_string` and `Sequence.from_bytes`
