### 2022-04-13 a390547
Reorganize code in `Sequence._shine_dalgarno_exact` and `Sequence._shine_dalgarno_mm`
