### 2022-04-13 243a540
Refactor Shine-Dalgarno methods to force Cython to generate C switches
