+++ pyrodigal/_pyrodigal.pyx
@@ -656,88 +656,70 @@ cdef class Sequence:
         # Compare the 6-base region to AGGAGG
         limit = min(6, start - 4 - pos)
         for i in range(limit):
-            if i%3 == 0 and _is_a(self.digits, self.slen, pos+i, strand):
-                match[i] = 2
-            elif i%3 != 0 and _is_g(self.digits, self.slen, pos+i, strand):
-                match[i] = 3
+            if i%3 == 0:
+                if _is_a(self.digits, self.slen, pos+i, strand):
+                    match[i] = 2
+            else:
+                if _is_g(self.digits, self.slen, pos+i, strand):
+                    match[i] = 3
 
         # Find the maximally scoring motif
         max_val = 0
         for i in range(limit, 2, -1):
             for j in range(limit+1-i):
+                # count number of matching positions, skip if less than
+                # GAG matching (or if a mismatch occurs)
                 cur_ctr = -2
-                mism = 0;
                 for k in range(j, j+i):
                     cur_ctr += match[k]
-                rdis = start - (pos + j + i)
-                if rdis > 15 or cur_ctr < 6:
+                if cur_ctr < 6:
                     continue
-
-                if rdis < 5 and i < 5:
-                    dis_flag = 2
-                elif rdis < 5 and i >= 5:
-                    dis_flag = 1
-                elif rdis > 10 and rdis <= 12 and i < 5:
-                    dis_flag = 1
-                elif rdis > 10 and rdis <= 12 and i >= 5:
-                    dis_flag = 2
-                elif rdis >= 13:
+                # compute distance to the start codon
+                rdis = start - (pos + j + i)
+                if rdis < 5: # 3-4bp
+                    dis_flag = 2 if i < 5 else 1
+                elif rdis < 11: # 5-10bp
+                    dis_flag = 0
+                elif rdis < 13: # 11-12bp
+                    dis_flag = 1 if i < 5 else 2
+                elif rdis < 16: # 13-15bp
                     dis_flag = 3
                 else:
-                    dis_flag = 0
-
-                # Exact-Matching RBS Motifs
-                if cur_ctr == 6 and dis_flag == 2:
-                    cur_val = 1
-                elif cur_ctr == 6 and dis_flag == 3:
-                    cur_val = 2
-                elif cur_ctr == 8 and dis_flag == 3:
-                    cur_val = 3
-                elif cur_ctr == 9 and dis_flag == 3:
-                    cur_val = 3
-                elif cur_ctr == 6 and dis_flag == 1:
-                    cur_val = 6
-                elif cur_ctr == 11 and dis_flag == 3:
-                    cur_val = 10
-                elif cur_ctr == 12 and dis_flag == 3:
-                    cur_val = 10
-                elif cur_ctr == 14 and dis_flag == 3:
-                    cur_val = 10
-                elif cur_ctr == 8 and dis_flag == 2:
-                    cur_val = 11
-                elif cur_ctr == 9 and dis_flag == 2:
-                    cur_val = 11
-                elif cur_ctr == 8 and dis_flag == 1:
-                    cur_val = 12
-                elif cur_ctr == 9 and dis_flag == 1:
-                    cur_val = 12
-                elif cur_ctr == 6 and dis_flag == 0:
-                    cur_val = 13
-                elif cur_ctr == 8 and dis_flag == 0:
-                    cur_val = 15
-                elif cur_ctr == 9 and dis_flag == 0:
-                    cur_val = 16
-                elif cur_ctr == 11 and dis_flag == 2:
-                    cur_val = 20
-                elif cur_ctr == 11 and dis_flag == 1:
-                    cur_val = 21
[…]
