### 2022-04-14 fbe7a81
Rewrite connection scoring into different functions depending on the target node
