+++ pyrodigal/_connection.h
@@ -1,0 +1,512 @@
+#ifndef _PYRODIGAL_CONNECTION_H
+#define _PYRODIGAL_CONNECTION_H
+
+#include <stdlib.h>
+#include "node.h"
+#include "training.h"
+
+static inline double _intergenic_mod_diff(
+    const struct _node* n1,
+    const struct _node* n2,
+    const double        start_weight
+) {
+    return -0.15 * start_weight;
+}
+
+
+static inline double _intergenic_mod_same(
+    const struct _node* n1,
+    const struct _node* n2,
+    const double        start_weight
+) {
+    int    dist    = abs(n1->ndx - n2->ndx);
+    int    overlap = n1->ndx + 2*n1->strand >= n2->ndx;
+    double rval    = 0.0;
+
+    if ((n1->ndx + 2 == n2->ndx) || (n1->ndx == n2->ndx + 1)) {
+        if (n1->strand == 1) {
+            if (n2->rscore < 0) rval -= n2->rscore;
+            if (n2->uscore < 0) rval -= n2->uscore;
+        } else {
+            if (n1->rscore < 0) rval -= n1->rscore;
+            if (n1->uscore < 0) rval -= n1->uscore;
+        }
+    }
+
+    if (dist > 3 * OPER_DIST) {
+        rval -= 0.15 * start_weight;
+    } else if (((dist <= OPER_DIST) && !overlap) || (dist * 4 < OPER_DIST)) {
+        rval += (2.0 - ((double) dist / OPER_DIST)) * 0.15 * start_weight;
+    }
+
+    return rval;
+}
+
+
+static inline double _intergenic_mod(
+    const struct _node* n1,
+    const struct _node* n2,
+    const double        start_weight
+) {
+    if (n1->strand == n2->strand) {
+        return _intergenic_mod_same(n1, n2, start_weight);
+    } else {
+        return _intergenic_mod_diff(n1, n2, start_weight);
+    }
+}
+
+
+static inline void _score_connection(
+    const struct _node*              nodes,
+    const struct _node*     restrict n1,
+          struct _node*     restrict n2,
+    const struct _training*          tinf,
+    const int                        final
+) {
+    const struct _node* n3;
+
+    int i;
+    int bnd;
+    int ovlp  = 0;
+    int maxfr = -1;
+    int left  = n1->ndx;
+    int right = n2->ndx;
+
+    double maxval;
+    double score   = 0.0;
+    double scr_mod = 0.0;
+
+    // NOTE(@althonos): We can skip checking for invalid connections here
+    //                  because the connection scorer already checked for
+    //                  them using the SIMD filter
+
+    // --- Edge Artifacts ---
+    if ((n1->traceb == -1) && (n1->strand == 1) && (n1->type == STOP)) {
+        return;
+    } else if ((n1->traceb == -1) && (n1->strand == -1) && (n1->type != STOP)) {
+        return;
+    }
+
+    // --- Genes ---
+    // 5'fwd->3'fwd
+    if ((n1->strand == 1) && (n2->strand == 1) && (n1->type != STOP) && (n2->type == STOP)) {
+        if (n2->stop_val >= n1->ndx)
+            return;
+        right += 2;
+        if (final)
+            score = n1->cscore + n1->sscore;
+        else
+            scr_mod = tinf->bias[0]*n1->gc_score[0] + tinf->bias[1]*n1->gc_score[1] + tinf->bias[2]*n1->gc_score[2];
+
+    // 3'rev->5'rev
+    } else if ((n1->strand == -1) && (n2->strand == -1) && (n1->type == STOP) && (n2->type != STOP)) {
+        if (n1->stop_val <= n2->ndx)
[…]
