+++ pyrodigal/_pyrodigal.pyx
@@ -82,6 +82,15 @@ from pyrodigal._sequence cimport (
     _letters,
     _complement
 )
+from pyrodigal._connection cimport (
+    _intergenic_mod_same,
+    _intergenic_mod_diff,
+    _intergenic_mod,
+    _score_connection_forward_start,
+    _score_connection_forward_stop,
+    _score_connection_backward_start,
+    _score_connection_backward_stop,
+)
 from pyrodigal.impl.generic cimport skippable_generic
 
 IF TARGET_CPU == "x86":
@@ -1097,7 +1106,7 @@ cdef class ConnectionScorer:
             if left >= right:
                 return
             if final:
-                score = Node._intergenic_mod_same(n1, n2, tinf.st_wt)
+                score = _intergenic_mod_same(n1, n2, tinf.st_wt)
         # 3'fwd->3'rev
         elif n1.strand == 1 and n2.strand == -1 and n1.type == node_type.STOP and n2.type == node_type.STOP:
             left += 2
@@ -1121,7 +1130,7 @@ cdef class ConnectionScorer:
                         continue
                     # record max frame
                     if final == 1:
-                        score = n3.cscore + n3.sscore + Node._intergenic_mod(n3, n2, tinf.st_wt)
+                        score = n3.cscore + n3.sscore + _intergenic_mod(n3, n2, tinf.st_wt)
                     else:
                         score = tinf.bias[0]*n3.gc_score[0] + tinf.bias[1]*n3.gc_score[1] + tinf.bias[2]*n3.gc_score[2]
                     if score > maxval:
@@ -1130,24 +1139,24 @@ cdef class ConnectionScorer:
             if maxfr != -1:
                 n3 = &nodes.nodes[n2.star_ptr[maxfr]]
                 if final:
-                    score = n3.cscore + n3.sscore + Node._intergenic_mod(n3, n2, tinf.st_wt)
+                    score = n3.cscore + n3.sscore + _intergenic_mod(n3, n2, tinf.st_wt)
                 else:
                     scr_mod = tinf.bias[0]*n3.gc_score[0] + tinf.bias[1]*n3.gc_score[1] + tinf.bias[2]*n3.gc_score[2]
             elif final:
-                score = Node._intergenic_mod_diff(n1, n2, tinf.st_wt)
+                score = _intergenic_mod_diff(n1, n2, tinf.st_wt)
         # 5'rev->3'rev
         elif n1.strand == -1 and n2.strand == -1 and n1.type != node_type.STOP and n2.type == node_type.STOP:
             right -= 2
             if left >= right:
                 return
             if final:
-                score = Node._intergenic_mod_same(n1, n2, tinf.st_wt)
+                score = _intergenic_mod_same(n1, n2, tinf.st_wt)
         # 5'rev->5'fwd
         elif n1.strand == -1 and n2.strand == 1 and n1.type != node_type.STOP and n2.type != node_type.STOP:
             if left >= right:
                 return
             if final:
-                score = Node._intergenic_mod_diff(n1, n2, tinf.st_wt)
+                score = _intergenic_mod_diff(n1, n2, tinf.st_wt)
 
         # --- Possible Operons */ ---
         # 3'fwd->3'fwd, check for a start just to left of first 3'
@@ -1160,7 +1169,7 @@ cdef class ConnectionScorer:
             left = n3.ndx
             right += 2
             if final:
-                score = n3.cscore + n3.sscore + Node._intergenic_mod(n1, n3, tinf.st_wt)
+                score = n3.cscore + n3.sscore + _intergenic_mod(n1, n3, tinf.st_wt)
             else:
[…]
