### 2022-04-14 #4296 @standage
The author describes Pyrodigal, a software package implementing a Python interface to the widely-adopted gene prediction program Prodigal. The Prodigal source code is bundled with Pyrodigal and compiled as a Python extension, which Pyrodigal accesses directly via a Python interface rather than indirect system calls to a separate executable. In addition to Python bindings, the author describes performance optimizations made to Prodigal’s core scoring procedure.

Prodigal is the *de facto* standard for prokaryotic gene prediction in the genome informatics community, and Python is the community’s *de facto* general-purpose scripting language. Exposing Prodigal’s internals comprehensively and efficiently to Python programs is a valuable contribution, as underscored by Pyrodigal's usage statistics and its integration into various published workflows. **I am happy to endorse this submission for publication at JOSS**.

- I was able to install Pyrodigal painlessly from both PyPI and Bioconda. The PyPI/pip package management system is much less heavyweight than the Conda/Bioconda system, so the author’s point about not requiring a separately handled external executable with pip installation is well taken.
- Source code is only one component of many in an open source project. The Pyrodigal repository also includes CI configuration, an active issue tracker, descriptions of the software (both extensive and detailed), example usages, and comprehensive API documentation. The quality of these resources is at the level one would expect for a mature open source project with numerous contributors.
- I was able to write a custom Python script with Pyrodigal in just a few minutes, and it performed as expected on data from a handful of bacterial isolates. Performance versus equivalent Prodigal invocations is consistent with the benchmarking reported in the submitted manuscript. I do not have convenient access to binned metagenomic assemblies for testing of those features.
- Pyrodigal performed well on a battery of tests designed to stress or break the code. Exception messages were easy to understand, with no verbose stack traces or cryptic error codes/messages.
- The manuscript is clearly written, and after multiple read-throughs I cannot find any significant errors or mistakes requiring revision.
