### 2022-04-14 da417ab
Move `_score_connections` into a C header and use function pointers where applicable
