+++ pyrodigal/_connection.h
@@ -2,8 +2,10 @@
 #define _PYRODIGAL_CONNECTION_H
 
 #include <stdlib.h>
+#include <stdint.h>
 #include "node.h"
 #include "training.h"
+
 
 static inline double _intergenic_mod_diff(
     const struct _node* n1,
@@ -52,180 +54,6 @@ static inline double _intergenic_mod(
         return _intergenic_mod_same(n1, n2, start_weight);
     } else {
         return _intergenic_mod_diff(n1, n2, start_weight);
-    }
-}
-
-
-static inline void _score_connection(
-    const struct _node*              nodes,
-    const struct _node*     restrict n1,
-          struct _node*     restrict n2,
-    const struct _training*          tinf,
-    const int                        final
-) {
-    const struct _node* n3;
-
-    int i;
-    int bnd;
-    int ovlp  = 0;
-    int maxfr = -1;
-    int left  = n1->ndx;
-    int right = n2->ndx;
-
-    double maxval;
-    double score   = 0.0;
-    double scr_mod = 0.0;
-
-    // NOTE(@althonos): We can skip checking for invalid connections here
-    //                  because the connection scorer already checked for
-    //                  them using the SIMD filter
-
-    // --- Edge Artifacts ---
-    if ((n1->traceb == -1) && (n1->strand == 1) && (n1->type == STOP)) {
-        return;
-    } else if ((n1->traceb == -1) && (n1->strand == -1) && (n1->type != STOP)) {
-        return;
-    }
-
-    // --- Genes ---
-    // 5'fwd->3'fwd
-    if ((n1->strand == 1) && (n2->strand == 1) && (n1->type != STOP) && (n2->type == STOP)) {
-        if (n2->stop_val >= n1->ndx)
-            return;
-        right += 2;
-        if (final)
-            score = n1->cscore + n1->sscore;
-        else
-            scr_mod = tinf->bias[0]*n1->gc_score[0] + tinf->bias[1]*n1->gc_score[1] + tinf->bias[2]*n1->gc_score[2];
-
-    // 3'rev->5'rev
-    } else if ((n1->strand == -1) && (n2->strand == -1) && (n1->type == STOP) && (n2->type != STOP)) {
-        if (n1->stop_val <= n2->ndx)
-            return;
-        left -= 2;
-        if (final)
-            score = n2->cscore + n2->sscore;
-        else
-            scr_mod = tinf->bias[0]*n2->gc_score[0] + tinf->bias[1]*n2->gc_score[1] + tinf->bias[2]*n2->gc_score[2];
-
-    // --- Intergenic space (Noncoding) ---
-    // 3'fwd->5'fwd
-    } else if ((n1->strand == 1) && (n2->strand == 1) && (n1->type == STOP) && (n2->type != STOP)) {
-        left += 2;
-        if (left >= right)
-            return;
-        if (final)
-            score = _intergenic_mod_same(n1, n2, tinf->st_wt);
-    // 3'fwd->3'rev
-    } else if ((n1->strand == 1) && (n2->strand == -1) && (n1->type == STOP) && (n2->type == STOP)) {
-        left += 2;
-        right -= 2;
-        if (left >= right)
-            return;
-        // Overlapping Gene Case 2: Three consecutive overlapping genes f r r
-        maxfr = -1;
-        maxval = 0.0;
-        if (n1->traceb != -1) {
-            for (i = 0; i < 3; i++) {
-                if (n2->star_ptr[i] == -1)
-                    continue;
-                n3 = &nodes[n2->star_ptr[i]];
-                ovlp = left - n3->stop_val + 3;
[…]
