+++ pyrodigal/_pyrodigal.pyx
@@ -90,6 +90,9 @@ from pyrodigal._connection cimport (
     _score_connection_forward_stop,
     _score_connection_backward_start,
     _score_connection_backward_stop,
+    _score_connections,
+    connection_function,
+    CONNECTION_FUNCTIONS,
 )
 from pyrodigal.impl.generic cimport skippable_generic
 
@@ -1046,175 +1049,7 @@ cdef class ConnectionScorer:
             return 0
         return 0
 
-    @staticmethod
-    cdef void _score_connection(
-        Nodes nodes,
-        const int p1,
-        const int p2,
-        const _training* tinf,
-        const bint final
-    ) nogil:
-        cdef _node* n1 = &nodes.nodes[p1]
-        cdef _node* n2 = &nodes.nodes[p2]
-        cdef _node* n3
-
-        cdef int i
-        cdef int bnd
-        cdef int ovlp  = 0
-        cdef int maxfr = -1
-        cdef int left  = n1.ndx
-        cdef int right = n2.ndx
-
-        cdef double maxval
-        cdef double score   = 0.0
-        cdef double scr_mod = 0.0
-
-        # NOTE(@althonos): We can skip checking for invalid connections here
-        #                  because the connection scorer already checked for
-        #                  them using the SIMD filter
-
-        # --- Edge Artifacts ---
-        if n1.traceb == -1 and n1.strand == 1 and n1.type == node_type.STOP:
-            return
-        elif n1.traceb == -1 and n1.strand == -1 and n1.type != node_type.STOP:
-            return
-
-        # --- Genes ---
-        # 5'fwd->3'fwd
-        elif n1.strand == 1 and n2.strand == 1 and n1.type != node_type.STOP and n2.type == node_type.STOP:
-            if n2.stop_val >= n1.ndx:
-                return
-            right += 2
-            if final:
-                score = n1.cscore + n1.sscore
-            else:
-                scr_mod = tinf.bias[0]*n1.gc_score[0] + tinf.bias[1]*n1.gc_score[1] + tinf.bias[2]*n1.gc_score[2]
-        # 3'rev->5'rev */
-        elif n1.strand == -1 and n2.strand == -1 and n1.type == node_type.STOP and n2.type != node_type.STOP:
-            if n1.stop_val <= n2.ndx:
-                return
-            left -= 2;
-            if final == 0:
-                scr_mod = tinf.bias[0]*n2.gc_score[0] + tinf.bias[1]*n2.gc_score[1] + tinf.bias[2]*n2.gc_score[2]
-            elif final:
-                score = n2.cscore + n2.sscore
-
-        # --- Intergenic space (Noncoding) ---
-        # 3'fwd->5'fwd
-        elif n1.strand == 1 and n2.strand == 1 and n1.type == node_type.STOP and n2.type != node_type.STOP:
-            left += 2
-            if left >= right:
-                return
-            if final:
-                score = _intergenic_mod_same(n1, n2, tinf.st_wt)
-        # 3'fwd->3'rev
-        elif n1.strand == 1 and n2.strand == -1 and n1.type == node_type.STOP and n2.type == node_type.STOP:
-            left += 2
-            right -= 2
-            if left >= right:
-                return
-            # Overlapping Gene Case 2: Three consecutive overlapping genes f r r
-            maxfr = -1
-            maxval = 0.0
-            if n1.traceb != -1:
-                for i in range(3):
-                    if n2.star_ptr[i] == -1:
-                        continue
-                    n3 = &nodes.nodes[n2.star_ptr[i]]
-                    ovlp = left - n3.stop_val + 3
-                    if ovlp <= 0 or ovlp >= dprog.MAX_OPP_OVLP:
-                        continue
[…]
