### 2022-04-14 90b51ea
Document methods of `ConnectionScorer` and add `__sizeof__` method
