### 2022-04-14 6c4efa6
Fix rendering of some docstrings in `pyrodigal.Gene` [ci skip]
