### 2022-04-15 79dacd3
Reorganize allocation code and use a smaller growth factor for dynamic arrays
