### 2022-04-17 #4296 @Gab0
### Overview

This project consists of Cython bindings to the widely used prokaryotic ORF finder software `Prodigal`.

The premise is interesting, allowing `Prodigal` functions to be called directly from Python code as an alternative to binary calls using `subprocess.call` or similar. This enables a more flexible pipeline design.

The performance claims over CPU time, RAM use and storage use can be useful when analyzing large sets of genomes, a scenario that has become increasingly common over the last years.

The paper includes all the required information. It also includes interesting details about performance optimizations done on top of the `Prodigal` codebase.

### Setup

The `README.md` describes the following installation methods:

* Python's standard package repository `pypi`.
* Bioconda.

From these I tested the `pypi` method and it works nicely, but I don't use `Bioconda`.

There seems to be two undocumented installation methods:

* The [AUR package](https://aur.archlinux.org/packages/python-pyrodigal) named `python-pyrodigal`.
* Build from the git source.

Both methods works nicely. The local build works after pulling all the submodules to the local repository.

### Tests

I wrote a quick comparison test for `Pyrodigal` against the `Prodigal` binary. The workflow is a basic run using default parameters. The `Pyrodigal` binary runs much faster (1.8x here) than the original `Prodigal` executable, reflecting the performance improvements described in the paper.

[ ] However, in this run, calling `Pyrodigal` routines using the library took more time than with the binary. Execution times are 15.5s against 2.5s. Maybe I did not understand the sample code shown at the `README.md` and wrote wrong library calls.

I would like to ask the author a few questions about these benchmark results:

1. Do both snippets shown below have equivalent functionality?

a) `pyrodigal -i <genome.fna> -o genes -a proteins.faa`

b)
```
orf_finder = pyrodigal.OrfFinder(meta=True)

record = SeqIO.read(GENOME, "fasta")
genome_input = record.seq.encode("utf-8")

predictions = orf_finder.find_genes(genome_input)
with open("proteins.faa", 'w', encoding="utf-8") as f:
    for i, gene in enumerate(predictions):
        f.write(f">{record.id}_{i+1}\n")
        f.write(gene.translate() + '\n')
```

2. If they are equivalent, is this performance difference expected? Maybe it is related to something in my setup. If applicable, explanations about performance differences between library and binary would be a good addition to the documentation.

### Suggestions

[ ] The package could be improved by including complete example workflows, possibly including sample data. The code snippets provided in the `README.md`
are basic demonstrations that require modification to be executed. They also include `import` statements that won't work with current package versions (e.g. `import Bio.SeqIO`, `import multiprocessing.pool`).

[ ] Automatic linters such as `autopep8` could be applied to the source code. At least for the pure python files, since I don't know if they are effective for Cython code. That should improve code readability by enforcing line length limits and
solving other minor formatting issues.

### Summary

1. The documentation and the paper are well written and contain all the relevant information.

2. Setup works nicely.

3. Expected performance gains are observable when comparing `Pyrodigal` executable against the original `Prodigal`.

4. I have a few questions about library performance in the `Tests` section. These correspond to the remaining checkboxes in my review.

The project is solid overall. My review status is `minor revision`.
