### 2022-04-18 d9e938a
Add a doc section on how to install from the Arch User Repository [ci skip]
