+++ pyrodigal/tests/test_genes.py
@@ -13,7 +13,6 @@ from .fasta import parse
 
 
 class TestGenes(unittest.TestCase):
-
     @classmethod
     def setUpClass(cls):
         data = os.path.realpath(os.path.join(__file__, "..", "data"))
@@ -32,14 +31,14 @@ class TestGenes(unittest.TestCase):
         with self.assertRaises(IndexError):
             self.genes[length]
         with self.assertRaises(IndexError):
-            self.genes[-length-1]
+            self.genes[-length - 1]
 
     def test_iter(self):
         for i, gene in zip(range(len(self.genes)), self.genes):
             self.assertEqual(gene._gene_data, self.genes[i]._gene_data)
 
     def test_reversed(self):
-        for i, gene in zip(range(1, len(self.genes)+1), reversed(self.genes)):
+        for i, gene in zip(range(1, len(self.genes) + 1), reversed(self.genes)):
             self.assertEqual(gene._gene_data, self.genes[-i]._gene_data)
 
     def test_bool(self):
@@ -51,7 +50,7 @@ class TestGenes(unittest.TestCase):
         for gene in self.genes:
             self.assertIn(gene.translation_table, valid)
 
-    @unittest.skipIf(sys.implementation.name != 'cpython', 'can panic with PyPy')
+    @unittest.skipIf(sys.implementation.name != "cpython", "can panic with PyPy")
     def test_collection_abc_subclass(self):
         self.assertIsInstance(self.genes, collections.abc.Sequence)
         self.assertIsInstance(self.genes, collections.abc.Sized)
@@ -66,18 +65,14 @@ class TestGenes(unittest.TestCase):
         actual = [
             line.strip()
             for line in buffer.getvalue().splitlines()
-            if not line.startswith("#")
-            and line.strip()
+            if not line.startswith("#") and line.strip()
         ]
 
         data = os.path.realpath(os.path.join(__file__, "..", "data"))
         tsv = os.path.join(data, "SRR492066.meta.tsv")
         with open(tsv) as f:
             expected = [
-                line.strip()
-                for line in f
-                if not line.startswith("#")
-                and line.strip()
+                line.strip() for line in f if not line.startswith("#") and line.strip()
             ]
 
         r1 = csv.reader(actual, dialect="excel-tab")
