+++ pyrodigal/tests/utils.py
@@ -10,11 +10,13 @@ def load_record(name):
     with gzip.open(fna, "rt") as f:
         return next(parse(f))
 
+
 def load_proteins(name, mode):
     data = os.path.realpath(os.path.join(__file__, "..", "data"))
     faa = os.path.join(data, "{name}.{mode}.faa.gz".format(name=name, mode=mode))
     with gzip.open(faa, "rt") as f:
         return list(parse(f))
+
 
 def load_genes(name, mode):
     data = os.path.realpath(os.path.join(__file__, "..", "data"))
