### 2022-04-18 #4296 @althonos
Hi @Gab0 , thanks for the review! A few answers / updates:

> There seems to be two undocumented installation methods:
>
>    The [AUR package](https://aur.archlinux.org/packages/python-pyrodigal) named python-pyrodigal.
>   Build from the git source.

I added the AUR installation method recently, that's why it's not documented, and updated the documentation to mention it, so it will be rendered on the website with the next release. Installing from source is not covered in the `README.md` (because it's really less convenient than PyPI / Bioconda), but I already have a section in the [Install page}(https://pyrodigal.readthedocs.io/en/stable/install.html) of the documentation.

> I wrote a quick comparison test for Pyrodigal against the Prodigal binary. The workflow is a basic run using default parameters. The Pyrodigal binary runs much faster (1.8x here) than the original Prodigal executable, reflecting the performance improvements described in the paper.
>
> However, in this run, calling Pyrodigal routines using the library took more time than with the binary. Execution times are 15.5s against 2.5s. Maybe I did not understand the sample code shown at the README.md and wrote wrong library calls.
>
> I would like to ask the author a few questions about these benchmark results:
>    Do both snippets shown below have equivalent functionality?
>
> If they are equivalent, is this performance difference expected? Maybe it is related to something in my setup. If applicable, explanations about performance differences between library and binary would be a good addition to the documentation.

No, the two snippets are not equivalent, because in the command line invocation you called Pyrodigal/Prodigal in _single_ mode (which trains and predicts on the same sequence) but in the script you used the _meta_ mode (as shown by the first line, `orf_finder = pyrodigal.OrfFinder(meta=True)`), which predicts genes using the best-scoring pre-trained data.

 > The package could be improved by including complete example workflows, possibly including sample data. The code snippets provided in the `README.md` are basic demonstrations that require modification to be executed. They also include import statements that won't work with current package versions (e.g. `import Bio.SeqIO`, `import multiprocessing.pool`).

Pyrodigal is not thought to perform I/O, which is why I included examples with [Biopython](https://biopython.org) and [Scikit-Bio](http://scikit-bio.org/), which are two general purpose libraries for bioinformatics, and can be used to load data from FASTA or GenBank files. `multiprocessing.pool` is a standard library module so it will work in any Python setup. I can however update the examples with paths to real data so that they are immediately testable.

> Automatic linters such as `autopep8` could be applied to the source code. At least for the pure python files, since I don't know if they are effective for Cython code. That should improve code readability by enforcing line length limits and
solving other minor formatting issues.

I have linted the Python sources with `black`, but to my knowledge there is no source code autoformatting tools for Cython. In addition, I'd rather _not_ reformat the Cython code, because some code blocks have been ported from the C code of Prodigal, and although they don't look very nice, having a 1-to-1 match between the Prodigal code and the Pyrodigal code for this sections improves readability when comparing the two, which also makes the code more future-proof if the Prodigal code changes in the future (because diffing will be easier).
