### 2022-04-18 #4296 @Gab0
@althonos Cool, I'll address each remaining topic:

* Extra Setup Methods: Setup using AUR and source builds do work and that was my main concern here.

* Library performance issues: Oh `meta` mode is another operation mode, got that. I'm now able to 'recreate' the standard `prodigal` operation workflow by training `pyrodigal.OrfFinder` on the same sequence it was going to predict. This workflow is actually included in the `README.md`, but wouldn't pointing out which example corresponds to the standard `prodigal` operation make things clearer for starter users? The performance with the library is as good as the `pyrodigal` binary, so there is no problem here.

* Example import issues: Let's forget this part, I messed up with the imports.

* Code linter: `black` fixed the issues with Python sources. Good point on leaving the Cython code as close as possible to `prodigal`.

Ok, so all my concerns have been addressed. My suggestions about working examples and documentation updates can be implemented in future releases if the author feels they will improve the package.

My review is completed, and I ****endorse this submission for publication at JOSS**. (@mikldk)
