### 2022-04-19 33348bc
Improve examples and Prodigal comparison in `README.md` [ci skip]
