### 2022-04-20 04f23d4
Fix remaining small issues with documentation
