+++ pyrodigal/_version.py
@@ -1,1 +1,1 @@
-__version__ = "0.7.3"
+__version__ = "1.0.0"
