+++ setup.cfg
@@ -11,7 +11,7 @@ license = GPLv3
 platform = any
 keywords = bioinformatics, genome, orf, gene, prodigal
 classifier =
-    Development Status :: 4 - Beta
+    Development Status :: 5 - Production/Stable
     Intended Audience :: Developers
     Intended Audience :: Science/Research
     License :: OSI Approved :: GNU General Public License v3 (GPLv3)
