### 2022-04-20 #4296 @althonos
Hi @mikldk ,

I have now made a new stable release (`v1.0.0`) and finalized the paper. The DOI for the `v1.0.0` version on Zenodo is [`10.5281/zenodo.6472583`](https://zenodo.org/record/6472583).
