### 2022-04-20 #4296 @althonos
> Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.

Does this means the archive needs to be named the same as the JOSS paper title?
