### 2022-04-25 #4296 @mikldk
@althonos Thanks. A few comments:

- [ ] There is a period in the end of the title (both paper and Zenodo) - please remove
- [ ] Summary: "One of the key step for analysing" -> step**s**

Please change and recompile (no need to make a new release, just update Zenodo meta data and the paper).
