### 2022-04-25 96dfc50
Fix title and grammatical error in `paper.md`
